Extract services CLI domain
