+++ crates/clickhousectl/src/cloud/cli.rs
@@ -1,8 +1,11 @@
 pub use crate::cloud::activity::ActivityCommands;
 pub use crate::cloud::api_keys::KeyCommands;
 pub use crate::cloud::auth::AuthCommands;
+#[allow(unused_imports)]
 pub use crate::cloud::backups::{BackupCommands, BackupConfigCommands};
 pub use crate::cloud::organizations::{InvitationCommands, MemberCommands, OrgCommands};
+#[allow(unused_imports)]
+pub use crate::cloud::services::{PrivateEndpointCommands, QueryEndpointCommands, ServiceCommands};
 use clap::builder::PossibleValuesParser;
 use clap::{Args, Subcommand};
 
@@ -220,29 +223,7 @@ impl CloudCommands {
         match self {
             CloudCommands::Auth { command } => command.is_write(),
             CloudCommands::Org { command } => command.is_write(),
-            CloudCommands::Service { command } => match command {
-                ServiceCommands::List { .. } => false,
-                ServiceCommands::Get { .. } => false,
-                ServiceCommands::Prometheus { .. } => false,
-                ServiceCommands::Query { .. } => false,
-                ServiceCommands::Create { .. } => true,
-                ServiceCommands::Delete { .. } => true,
-                ServiceCommands::Start { .. } => true,
-                ServiceCommands::Stop { .. } => true,
-                ServiceCommands::Update { .. } => true,
-                ServiceCommands::Scale { .. } => true,
-                ServiceCommands::ResetPassword { .. } => true,
-                ServiceCommands::QueryEndpoint { command } => match command {
-                    QueryEndpointCommands::Get { .. } => false,
-                    QueryEndpointCommands::Create { .. } => true,
-                    QueryEndpointCommands::Delete { .. } => true,
-                },
-                ServiceCommands::PrivateEndpoint { command } => match command {
-                    PrivateEndpointCommands::Create { .. } => true,
-                    PrivateEndpointCommands::GetConfig { .. } => false,
-                },
-                ServiceCommands::BackupConfig { command } => command.is_write(),
-            },
+            CloudCommands::Service { command } => command.is_write(),
             CloudCommands::Backup { command } => command.is_write(),
             CloudCommands::Member { command } => command.is_write(),
             CloudCommands::Invitation { command } => command.is_write(),
@@ -270,514 +251,6 @@ impl CloudCommands {
             },
         }
     }
-}
-
-#[derive(Subcommand)]
-pub enum ServiceCommands {
-    /// List all services
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Lists all services in the organization. Org ID is auto-detected if only one org exists.
-  Returns service IDs needed by get, delete, start, stop, and backup commands.
-  Add --json for machine-readable output.
-  Related: `clickhousectl cloud service get <id>` for full details.")]
-    List {
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-
-        /// Filter by resource tags (e.g., "tag:env=production")
-        #[arg(long)]
-        filter: Vec<String>,
-    },
-
-    /// Get service details
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Returns full service details: status, endpoints, scaling config, IP access list.
-  Get the service ID from `clickhousectl cloud service list`.
-  Add --json for machine-readable output.
-  Related: `clickhousectl cloud service start/stop <id>` to change state.")]
-    Get {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Create a new service
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Creates a new ClickHouse Cloud service. Only --name is required; other fields have defaults.
-  Returns the new service ID and initial password — save these.
-  Typical: `clickhousectl cloud service create --name my-svc`.
-  Defaults: provider=aws, region=us-east-1. Add --json for machine-readable output.
-  Related: `clickhousectl cloud service get <id>` to check status after creation.")]
-    Create {
-        /// Service name (required)
-        #[arg(long)]
-        name: String,
-
-        /// Cloud provider: aws, gcp, azure (required)
-        #[arg(long, default_value = "aws")]
-        provider: String,
-
-        /// Region (required). Examples: us-east-1, eu-west-1, us-central1
-        #[arg(long, default_value = "us-east-1")]
-        region: String,
-
-        /// Minimum memory per replica in GB (8-356, multiple of 4). Horizontal
-        /// autoscaling requires it equal to --max-replica-memory-gb.
-        #[arg(long)]
-        min_replica_memory_gb: Option<u32>,
-
-        /// Maximum memory per replica in GB (8-356, multiple of 4). Horizontal
-        /// autoscaling requires it equal to --min-replica-memory-gb.
-        #[arg(long)]
-        max_replica_memory_gb: Option<u32>,
-
-        /// Number of replicas (1-20). Vertical autoscaling; mutually exclusive
-        /// with the horizontal band (--min-replicas/--max-replicas).
-        #[arg(long, conflicts_with_all = ["min_replicas", "max_replicas"])]
-        num_replicas: Option<u32>,
-
-        /// Minimum number of replicas for horizontal autoscaling (requires the
-        /// horizontal autoscaling org feature). Mutually exclusive with --num-replicas.
-        #[arg(long, conflicts_with = "num_replicas")]
-        min_replicas: Option<u32>,
-
-        /// Maximum number of replicas for horizontal autoscaling (requires the
-        /// horizontal autoscaling org feature). Mutually exclusive with --num-replicas.
-        #[arg(long, conflicts_with = "num_replicas")]
-        max_replicas: Option<u32>,
-
-        /// Autoscaling mode: vertical (default) or horizontal. Horizontal uses fixed
-        /// memory per replica (--min-replica-memory-gb equal to --max-replica-memory-gb)
-        /// with a variable replica count (--min-replicas/--max-replicas); vertical uses
-        /// a fixed replica count (--num-replicas) with variable memory.
-        #[arg(
-            long,
-            value_parser = PossibleValuesParser::new(
-                clickhouse_cloud_api::models::AutoscalingMode::VALUES
-            )
-        )]
-        autoscaling_mode: Option<String>,
-
-        /// Allow scale to zero when idle (default: true)
-        #[arg(long)]
-        idle_scaling: Option<bool>,
-
-        /// Minimum idle timeout in minutes (>= 5)
-        #[arg(long)]
-        idle_timeout_minutes: Option<u32>,
-
-        /// IP addresses to allow (CIDR format, e.g., "0.0.0.0/0"). Can be specified multiple times
-        #[arg(long = "ip-allow")]
-        ip_allow: Vec<String>,
-
-        /// Backup ID to restore from
-        #[arg(long)]
-        backup_id: Option<String>,
-
-        /// Release channel: slow, default, fast
-        #[arg(long)]
-        release_channel: Option<String>,
-
-        /// Data warehouse ID (for creating read replicas)
-        #[arg(long)]
-        data_warehouse_id: Option<String>,
-
-        /// Make service read-only (requires --data-warehouse-id)
-        #[arg(long)]
-        readonly: bool,
-
-        /// Customer-provided disk encryption key
-        #[arg(long)]
-        encryption_key: Option<String>,
-
-        /// Role ARN for disk encryption
-        #[arg(long)]
-        encryption_role: Option<String>,
-
-        /// Enable Transparent Data Encryption (enterprise only)
-        #[arg(long)]
-        enable_tde: bool,
-
-        /// Compliance type: hipaa, pci
-        #[arg(long)]
-        compliance_type: Option<String>,
-
-        /// Instance profile (enterprise only): v1-default, v1-highmem-xs, etc.
-        #[arg(long)]
-        profile: Option<String>,
-
-        /// Tag to attach to the service. Format: key or key=value
-        #[arg(long = "tag", value_name = "KEY[=VALUE]")]
-        tag: Vec<String>,
-
-        /// Enable a toggleable endpoint protocol. Currently supported: mysql
-        #[arg(long = "enable-endpoint")]
-        enable_endpoint: Vec<String>,
-
-        /// Disable a toggleable endpoint protocol. Currently supported: mysql
-        #[arg(long = "disable-endpoint")]
-        disable_endpoint: Vec<String>,
-
-        /// Accept private preview terms for eligible service creation flows
-        #[arg(long)]
-        private_preview_terms_checked: bool,
-
-        /// Enable or disable service core dump collection
-        #[arg(long)]
-        enable_core_dumps: Option<bool>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Delete a service
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Permanently deletes a ClickHouse Cloud service. This action is irreversible.
-  Use --force to stop a running service before deleting it in one step.
-  Related: `clickhousectl cloud service stop <id>` to idle instead of delete.")]
-    Delete {
-        /// Service ID
-        service_id: String,
-
-        /// Stop the service first if it is running, then delete
-        #[arg(long)]
-        force: bool,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Start a service
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Resumes a stopped/idled ClickHouse Cloud service.
-  Takes a service ID — get it from `clickhousectl cloud service list`.
-  Add --json for machine-readable output.
-  Related: `clickhousectl cloud service get <id>` to check status, `clickhousectl cloud service stop <id>` to idle.")]
-    Start {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Stop a service
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Idles a ClickHouse Cloud service, stopping billing for compute.
-  Data is preserved. Takes a service ID — get it from `clickhousectl cloud service list`.
-  Add --json for machine-readable output.
-  Related: `clickhousectl cloud service start <id>` to resume, `clickhousectl cloud service delete <id>` to remove.")]
-    Stop {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Update service settings
-    Update {
-        /// Service ID
-        service_id: String,
-
-        /// New service name
-        #[arg(long)]
-        name: Option<String>,
-
-        /// Add an IP/CIDR entry to the service allow list
-        #[arg(long = "add-ip-allow")]
-        add_ip_allow: Vec<String>,
-
-        /// Remove an IP/CIDR entry from the service allow list
-        #[arg(long = "remove-ip-allow")]
-        remove_ip_allow: Vec<String>,
-
-        /// Add a private endpoint ID to the service
-        #[arg(long = "add-private-endpoint-id")]
-        add_private_endpoint_id: Vec<String>,
-
-        /// Remove a private endpoint ID from the service
-        #[arg(long = "remove-private-endpoint-id")]
-        remove_private_endpoint_id: Vec<String>,
-
-        /// Release channel: slow, default, fast
-        #[arg(long)]
-        release_channel: Option<String>,
-
-        /// Enable a toggleable endpoint protocol. Currently supported: mysql
-        #[arg(long = "enable-endpoint")]
-        enable_endpoint: Vec<String>,
-
-        /// Disable a toggleable endpoint protocol. Currently supported: mysql
-        #[arg(long = "disable-endpoint")]
-        disable_endpoint: Vec<String>,
-
-        /// Transparent Data Encryption key ID to rotate to
-        #[arg(long)]
-        transparent_data_encryption_key_id: Option<String>,
-
-        /// Tag to add. Format: key or key=value
-        #[arg(long = "add-tag", value_name = "KEY[=VALUE]")]
-        add_tag: Vec<String>,
-
-        /// Tag to remove. Format: key or key=value
-        #[arg(long = "remove-tag", value_name = "KEY[=VALUE]")]
-        remove_tag: Vec<String>,
-
-        /// Enable or disable service core dump collection
-        #[arg(long)]
-        enable_core_dumps: Option<bool>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Update replica scaling settings
-    Scale {
-        /// Service ID
-        service_id: String,
-
-        /// Minimum memory per replica in GB (8-356, multiple of 4). Horizontal
-        /// autoscaling requires it equal to --max-replica-memory-gb.
-        #[arg(long)]
-        min_replica_memory_gb: Option<u32>,
-
-        /// Maximum memory per replica in GB (8-356, multiple of 4). Horizontal
-        /// autoscaling requires it equal to --min-replica-memory-gb.
-        #[arg(long)]
-        max_replica_memory_gb: Option<u32>,
-
-        /// Number of replicas (1-20). Vertical autoscaling; mutually exclusive
-        /// with the horizontal band (--min-replicas/--max-replicas).
-        #[arg(long, conflicts_with_all = ["min_replicas", "max_replicas"])]
-        num_replicas: Option<u32>,
-
-        /// Minimum number of replicas for horizontal autoscaling (requires the
-        /// horizontal autoscaling org feature). Mutually exclusive with --num-replicas.
-        #[arg(long, conflicts_with = "num_replicas")]
-        min_replicas: Option<u32>,
-
-        /// Maximum number of replicas for horizontal autoscaling (requires the
-        /// horizontal autoscaling org feature). Mutually exclusive with --num-replicas.
-        #[arg(long, conflicts_with = "num_replicas")]
-        max_replicas: Option<u32>,
-
-        /// Autoscaling mode: vertical (default) or horizontal. Omit to keep the
-        /// service's current mode. See `service create --autoscaling-mode`.
-        #[arg(
-            long,
-            value_parser = PossibleValuesParser::new(
-                clickhouse_cloud_api::models::AutoscalingMode::VALUES
-            )
-        )]
-        autoscaling_mode: Option<String>,
-
-        /// Allow scale to zero when idle
-        #[arg(long)]
-        idle_scaling: Option<bool>,
-
-        /// Minimum idle timeout in minutes (>= 5)
-        #[arg(long)]
-        idle_timeout_minutes: Option<u32>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Reset a service's default user password
-    ResetPassword {
-        /// Service ID
-        service_id: String,
-
-        /// SHA256 password hash encoded as base64
-        #[arg(long)]
-        new_password_hash: Option<String>,
-
-        /// MySQL-compatible double SHA1 password hash
-        #[arg(long)]
-        new_double_sha1_hash: Option<String>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Manage query endpoints
-    #[command(name = "query-endpoint")]
-    QueryEndpoint {
-        #[command(subcommand)]
-        command: QueryEndpointCommands,
-    },
-
-    /// Manage private endpoints for a service
-    #[command(name = "private-endpoint")]
-    PrivateEndpoint {
-        #[command(subcommand)]
-        command: PrivateEndpointCommands,
-    },
-
-    /// Manage backup configuration for a service
-    #[command(name = "backup-config")]
-    BackupConfig {
-        #[command(subcommand)]
-        command: BackupConfigCommands,
-    },
-
-    /// Get service Prometheus metrics
-    Prometheus {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-
-        /// Whether to request filtered metrics
-        #[arg(long)]
-        filtered_metrics: Option<bool>,
-    },
-
-    /// Run a SQL query against a cloud service over HTTP via the Query API
-    #[command(after_help = "\
-CONTEXT FOR AGENTS:
-  Runs SQL over HTTP — no local clickhouse binary or service password required.
-  With API key auth: first uses the authenticated key directly when the query
-  endpoint already authorizes it. Otherwise, a per-service read+write key is
-  auto-provisioned and stored in .clickhouse/credentials.json.
-  With OAuth (cloud auth login): sends your own bearer token — SQL runs as
-  your cloud user with read-only access (SELECT only, no writes); no key
-  provisioning and no query endpoint required on the service.
-  SQL precedence: --query > --queries-file > stdin. Default format: PrettyCompact
-  on a TTY, TabSeparated when piped. --json selects JSONEachRow and cannot be
-  combined with --format; an explicit --format takes precedence over agent
-  auto-JSON.")]
-    Query {
-        /// Service name to query (exactly one of --name or --id is required)
-        #[arg(long, conflicts_with = "id")]
-        name: Option<String>,
-
-        /// Service ID to query
-        #[arg(long, conflicts_with = "name")]
-        id: Option<String>,
-
-        /// Execute a SQL query
-        #[arg(long, short)]
-        query: Option<String>,
-
-        /// Execute queries from a SQL file (use "-" for stdin)
-        #[arg(long)]
-        queries_file: Option<String>,
-
-        /// Target database
-        #[arg(long)]
-        database: Option<String>,
-
-        /// Response format (e.g. JSONEachRow, CSV, TabSeparated, PrettyCompact)
-        #[arg(long, conflicts_with = "json")]
-        format: Option<String>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-
-        /// Fail instead of auto-provisioning when the authenticated API key
-        /// cannot use the query endpoint and no key is stored locally (API
-        /// key auth only; with OAuth this flag has no effect)
-        #[arg(long)]
-        no_auto_enable: bool,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum QueryEndpointCommands {
-    /// Get query endpoint configuration
-    Get {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Create/enable query endpoint
-    Create {
-        /// Service ID
-        service_id: String,
-
-        /// Roles to grant access (can be specified multiple times)
-        #[arg(long)]
-        role: Vec<String>,
-
-        /// OpenAPI key IDs to authorize
-        #[arg(long = "open-api-key")]
-        open_api_key: Vec<String>,
-
-        /// Allowed origins string for browser access
-        #[arg(long)]
-        allowed_origins: Option<String>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Delete/disable query endpoint
-    Delete {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum PrivateEndpointCommands {
-    /// Create a private endpoint connection
-    Create {
-        /// Service ID
-        service_id: String,
-
-        /// Private endpoint ID (VPC endpoint ID)
-        #[arg(long)]
-        endpoint_id: String,
-
-        /// Description
-        #[arg(long)]
-        description: Option<String>,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
-
-    /// Get service private endpoint configuration
-    GetConfig {
-        /// Service ID
-        service_id: String,
-
-        /// Organization ID (auto-detected if not specified)
-        #[arg(long)]
-        org_id: Option<String>,
-    },
 }
 
 #[derive(Subcommand)]
@@ -1591,380 +1064,6 @@ mod tests {
     use clap::Parser;
 
     #[test]
-    fn parses_service_query_json_mode() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "query",
-            "--id",
-            "svc-1",
-            "--query",
-            "SELECT 1",
-            "--json",
-        ])
-        .unwrap();
-
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        assert!(args.json);
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Query { format, .. } = command else {
-            panic!("expected service query");
-        };
-        assert!(format.is_none());
-    }
-
-    #[test]
-    fn rejects_service_query_json_with_explicit_format() {
-        let error = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "query",
-            "--id",
-            "svc-1",
-            "--query",
-            "SELECT 1",
-            "--json",
-            "--format",
-            "CSV",
-        ])
-        .err()
-        .expect("--json and --format should conflict");
-
-        assert_eq!(error.kind(), clap::error::ErrorKind::ArgumentConflict);
-        let rendered = error.to_string();
-        assert!(rendered.contains("--json"));
-        assert!(rendered.contains("--format"));
-    }
-
-    #[test]
-    fn parses_service_query_explicit_format_without_json() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "query",
-            "--id",
-            "svc-1",
-            "--query",
-            "SELECT 1",
-            "--format",
-            "CSV",
-        ])
-        .unwrap();
-
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        assert!(!args.json);
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Query { format, .. } = command else {
-            panic!("expected service query");
-        };
-        assert_eq!(format.as_deref(), Some("CSV"));
-    }
-
-    #[test]
-    fn parses_service_update_ga_patch_flags() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "update",
-            "svc-1",
-            "--add-ip-allow",
-            "10.0.0.0/8",
-            "--remove-ip-allow",
-            "0.0.0.0/0",
-            "--add-private-endpoint-id",
-            "pe-1",
-            "--remove-private-endpoint-id",
-            "pe-2",
-            "--release-channel",
-            "fast",
-            "--enable-endpoint",
-            "mysql",
-            "--add-tag",
-            "env=prod",
-            "--enable-core-dumps",
-            "true",
-        ])
-        .unwrap();
-
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Update {
-            service_id,
-            add_ip_allow,
-            remove_ip_allow,
-            add_private_endpoint_id,
-            remove_private_endpoint_id,
-            release_channel,
-            enable_endpoint,
-            add_tag,
-            enable_core_dumps,
-            ..
-        } = command
-        else {
-            panic!("expected service update");
-        };
-
-        assert_eq!(service_id, "svc-1");
-        assert_eq!(add_ip_allow, vec!["10.0.0.0/8"]);
-        assert_eq!(remove_ip_allow, vec!["0.0.0.0/0"]);
-        assert_eq!(add_private_endpoint_id, vec!["pe-1"]);
-        assert_eq!(remove_private_endpoint_id, vec!["pe-2"]);
-        assert_eq!(release_channel.as_deref(), Some("fast"));
-        assert_eq!(enable_endpoint, vec!["mysql"]);
-        assert_eq!(add_tag, vec!["env=prod"]);
-        assert_eq!(enable_core_dumps, Some(true));
-    }
-
-    #[test]
-    fn parses_service_create_horizontal_autoscaling_flags() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "create",
-            "--name",
-            "s",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-            "--autoscaling-mode",
-            "horizontal",
-        ])
-        .unwrap();
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Create {
-            min_replicas,
-            max_replicas,
-            autoscaling_mode,
-            num_replicas,
-            min_replica_memory_gb,
-            max_replica_memory_gb,
-            ..
-        } = command
-        else {
-            panic!("expected service create");
-        };
-        assert_eq!(min_replicas, Some(2));
-        assert_eq!(max_replicas, Some(8));
-        assert_eq!(autoscaling_mode.as_deref(), Some("horizontal"));
-        assert!(num_replicas.is_none());
-        assert!(min_replica_memory_gb.is_none());
-        assert!(max_replica_memory_gb.is_none());
-    }
-
-    #[test]
-    fn rejects_service_create_horizontal_vertical_mix() {
-        // --min-replicas conflicts with the vertical flags
-        let result = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "create",
-            "--name",
-            "s",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-            "--num-replicas",
-            "3",
-        ]);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parses_service_create_horizontal_mode_with_memory_bounds() {
-        // Horizontal requires equal memory bounds, so the memory flags must
-        // combine with the mode and the replica band in one invocation.
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "create",
-            "--name",
-            "s",
-            "--autoscaling-mode",
-            "horizontal",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-            "--min-replica-memory-gb",
-            "16",
-            "--max-replica-memory-gb",
-            "16",
-        ])
-        .unwrap();
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Create {
-            min_replicas,
-            max_replicas,
-            autoscaling_mode,
-            min_replica_memory_gb,
-            max_replica_memory_gb,
-            ..
-        } = command
-        else {
-            panic!("expected service create");
-        };
-        assert_eq!(min_replicas, Some(2));
-        assert_eq!(max_replicas, Some(8));
-        assert_eq!(autoscaling_mode.as_deref(), Some("horizontal"));
-        assert_eq!(min_replica_memory_gb, Some(16));
-        assert_eq!(max_replica_memory_gb, Some(16));
-    }
-
-    #[test]
-    fn rejects_service_create_invalid_autoscaling_mode() {
-        let result = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "create",
-            "--name",
-            "s",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-            "--autoscaling-mode",
-            "turbo",
-        ]);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parses_service_scale_horizontal_autoscaling_flags() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "scale",
-            "svc-1",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-            "--autoscaling-mode",
-            "horizontal",
-        ])
-        .unwrap();
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Scale {
-            service_id,
-            min_replicas,
-            max_replicas,
-            autoscaling_mode,
-            num_replicas,
-            ..
-        } = command
-        else {
-            panic!("expected service scale");
-        };
-        assert_eq!(service_id, "svc-1");
-        assert_eq!(min_replicas, Some(2));
-        assert_eq!(max_replicas, Some(8));
-        assert_eq!(autoscaling_mode.as_deref(), Some("horizontal"));
-        assert!(num_replicas.is_none());
-    }
-
-    #[test]
-    fn parses_service_scale_switch_to_vertical_in_one_call() {
-        // Switching a horizontal service back to vertical sends the mode and
-        // the target replica count in a single request.
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "scale",
-            "svc-1",
-            "--autoscaling-mode",
-            "vertical",
-            "--num-replicas",
-            "3",
-            "--min-replica-memory-gb",
-            "8",
-            "--max-replica-memory-gb",
-            "32",
-        ])
-        .unwrap();
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::Scale {
-            autoscaling_mode,
-            num_replicas,
-            min_replica_memory_gb,
-            max_replica_memory_gb,
-            min_replicas,
-            max_replicas,
-            ..
-        } = command
-        else {
-            panic!("expected service scale");
-        };
-        assert_eq!(autoscaling_mode.as_deref(), Some("vertical"));
-        assert_eq!(num_replicas, Some(3));
-        assert_eq!(min_replica_memory_gb, Some(8));
-        assert_eq!(max_replica_memory_gb, Some(32));
-        assert!(min_replicas.is_none());
-        assert!(max_replicas.is_none());
-    }
-
-    #[test]
-    fn rejects_service_scale_num_replicas_with_replica_band() {
-        let result = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "scale",
-            "svc-1",
-            "--num-replicas",
-            "3",
-            "--min-replicas",
-            "2",
-            "--max-replicas",
-            "8",
-        ]);
-        assert!(result.is_err());
-    }
-
-    #[test]
     fn parses_clickpipe_object_storage_ingestion_control_flags() {
         let cli = Cli::try_parse_from([
             "clickhousectl",
@@ -2189,62 +1288,6 @@ mod tests {
         assert!(matches!(command, ClickPipeSchemaDiscoverCommands::Kafka(_)));
     }
 
-    #[test]
-    fn parses_private_endpoint_config_and_password_hash_flags() {
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "reset-password",
-            "svc-1",
-            "--new-password-hash",
-            "sha256",
-            "--new-double-sha1-hash",
-            "sha1",
-        ])
-        .unwrap();
-
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::ResetPassword {
-            new_password_hash,
-            new_double_sha1_hash,
-            ..
-        } = command
-        else {
-            panic!("expected reset-password");
-        };
-        assert_eq!(new_password_hash.as_deref(), Some("sha256"));
-        assert_eq!(new_double_sha1_hash.as_deref(), Some("sha1"));
-
-        let cli = Cli::try_parse_from([
-            "clickhousectl",
-            "cloud",
-            "service",
-            "private-endpoint",
-            "get-config",
-            "svc-1",
-        ])
-        .unwrap();
-        let Commands::Cloud(args) = cli.command else {
-            panic!("expected cloud command");
-        };
-        let CloudCommands::Service { command } = args.command else {
-            panic!("expected service command");
-        };
-        let ServiceCommands::PrivateEndpoint { command } = command else {
-            panic!("expected private-endpoint command");
-        };
-        let PrivateEndpointCommands::GetConfig { service_id, .. } = command else {
-            panic!("expected get-config");
-        };
-        assert_eq!(service_id, "svc-1");
-    }
-
     /// Helper to assert a command parsed from CLI args is classified correctly.
     fn assert_write(args: &[&str], expected: bool) {
         let cli = Cli::try_parse_from(args).unwrap();
@@ -2261,17 +1304,6 @@ mod tests {
 
     #[test]
     fn is_write_command_read_only_commands() {
-        // Service reads
-        assert_write(&["clickhousectl", "cloud", "service", "list"], false);
-        assert_write(
-            &["clickhousectl", "cloud", "service", "get", "svc-1"],
-            false,
-        );
-        assert_write(
-            &["clickhousectl", "cloud", "service", "prometheus", "svc-1"],
-            false,
-        );
-
         // Backup reads
         assert_write(
             &["clickhousectl", "cloud", "backup", "list", "svc-1"],
@@ -2279,19 +1311,6 @@ mod tests {
         );
         assert_write(
             &["clickhousectl", "cloud", "backup", "get", "svc-1", "bk-1"],
-            false,
-        );
-
-        // Backup config read
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "backup-config",
-                "get",
-                "svc-1",
-            ],
             false,
         );
 
@@ -2303,32 +1322,6 @@ mod tests {
         assert_write(&["clickhousectl", "cloud", "activity", "list"], false);
         assert_write(
             &["clickhousectl", "cloud", "activity", "get", "act-1"],
-            false,
-        );
-
-        // Query endpoint read
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "query-endpoint",
-                "get",
-                "svc-1",
-            ],
-            false,
-        );
-
-        // Private endpoint read
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "private-endpoint",
-                "get-config",
-                "svc-1",
-            ],
             false,
         );
 
@@ -2377,84 +1370,6 @@ mod tests {
             true,
         );
 
-        // Service writes
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "create",
-                "--name",
-                "s",
-                "--provider",
-                "aws",
-                "--region",
-                "us-east-1",
-            ],
-            true,
-        );
-        assert_write(
-            &["clickhousectl", "cloud", "service", "delete", "svc-1"],
-            true,
-        );
-        assert_write(
-            &["clickhousectl", "cloud", "service", "start", "svc-1"],
-            true,
-        );
-        assert_write(
-            &["clickhousectl", "cloud", "service", "stop", "svc-1"],
-            true,
-        );
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "update",
-                "svc-1",
-                "--name",
-                "new",
-            ],
-            true,
-        );
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "scale",
-                "svc-1",
-                "--num-replicas",
-                "2",
-            ],
-            true,
-        );
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "reset-password",
-                "svc-1",
-            ],
-            true,
-        );
-
-        // Backup config write
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "backup-config",
-                "update",
-                "svc-1",
-                "--backup-period-hours",
-                "12",
-            ],
-            true,
-        );
-
         // Key writes
         assert_write(
             &["clickhousectl", "cloud", "key", "create", "--name", "k"],
@@ -2473,45 +1388,6 @@ mod tests {
             true,
         );
         assert_write(&["clickhousectl", "cloud", "key", "delete", "key-1"], true);
-
-        // Query endpoint writes
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "query-endpoint",
-                "create",
-                "svc-1",
-            ],
-            true,
-        );
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "query-endpoint",
-                "delete",
-                "svc-1",
-            ],
-            true,
-        );
-
-        // Private endpoint write
-        assert_write(
-            &[
-                "clickhousectl",
-                "cloud",
-                "service",
-                "private-endpoint",
-                "create",
-                "svc-1",
-                "--endpoint-id",
-                "ep-1",
-            ],
-            true,
-        );
 
         // Postgres writes
         assert_write(
