+++ crates/clickhousectl/src/cloud/client.rs
@@ -450,240 +450,6 @@ impl CloudClient {
         }
     }
 
-    // Service endpoints (delegated to library client)
-    pub async fn list_services(
-        &self,
-        org_id: &str,
-    ) -> Result<Vec<clickhouse_cloud_api::models::Service>> {
-        let response = self
-            .api()
-            .instance_get_list(org_id, &[])
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn list_services_filtered(
-        &self,
-        org_id: &str,
-        filters: &[String],
-    ) -> Result<Vec<clickhouse_cloud_api::models::Service>> {
-        let filter_refs: Vec<&str> = filters.iter().map(|s| s.as_str()).collect();
-        let response = self
-            .api()
-            .instance_get_list(org_id, &filter_refs)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_service(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::Service> {
-        let response = self
-            .api()
-            .instance_get(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_service_if_exists(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<Option<clickhouse_cloud_api::models::Service>> {
-        match self.api().instance_get(org_id, service_id).await {
-            Ok(response) => Self::unwrap_response(response).map(Some),
-            Err(clickhouse_cloud_api::Error::Api { status: 404, .. }) => Ok(None),
-            Err(error) => Err(self.convert_error_for_organization(error, org_id)),
-        }
-    }
-
-    pub async fn create_service(
-        &self,
-        org_id: &str,
-        request: &clickhouse_cloud_api::models::ServicePostRequest,
-    ) -> Result<clickhouse_cloud_api::models::ServicePostResponse> {
-        let response = self
-            .api()
-            .instance_create(org_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn delete_service_if_exists(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<Option<DeleteResponse>> {
-        match self.api().instance_delete(org_id, service_id).await {
-            Ok(response) => Ok(Some(DeleteResponse {
-                status: response.status,
-                request_id: response.request_id,
-            })),
-            Err(clickhouse_cloud_api::Error::Api { status: 404, .. }) => {
-                // The service and its organization share the same 404 shape.
-                // Confirm the request scope before declaring the service absent.
-                self.get_organization(org_id).await?;
-                Ok(None)
-            }
-            Err(error) => Err(self.convert_error_for_organization(error, org_id)),
-        }
-    }
-
-    pub async fn change_service_state(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        command: clickhouse_cloud_api::models::ServiceStatePatchRequestCommand,
-    ) -> Result<clickhouse_cloud_api::models::Service> {
-        use clickhouse_cloud_api::models::ServiceStatePatchRequest;
-        let request = ServiceStatePatchRequest {
-            command: Some(command),
-        };
-        let response = self
-            .api()
-            .instance_state_update(org_id, service_id, &request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Update service
-    pub async fn update_service(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::ServicePatchRequest,
-    ) -> Result<clickhouse_cloud_api::models::Service> {
-        let response = self
-            .api()
-            .instance_update(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Replica scaling
-    pub async fn update_replica_scaling(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::ServiceReplicaScalingPatchRequest,
-    ) -> Result<clickhouse_cloud_api::models::ServiceScalingPatchResponse> {
-        let response = self
-            .api()
-            .instance_replica_scaling_update(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Reset password
-    pub async fn reset_password(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::ServicePasswordPatchRequest,
-    ) -> Result<clickhouse_cloud_api::models::ServicePasswordPatchResponse> {
-        let response = self
-            .api()
-            .instance_password_update(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Query endpoint (delegated to library client)
-    pub async fn get_query_endpoint(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::ServiceQueryAPIEndpoint> {
-        let response = self
-            .api()
-            .instance_query_endpoint_get(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn create_query_endpoint(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::InstanceServiceQueryApiEndpointsPostRequest,
-    ) -> Result<clickhouse_cloud_api::models::ServiceQueryAPIEndpoint> {
-        let response = self
-            .api()
-            .instance_query_endpoint_upsert(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn delete_query_endpoint(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<DeleteResponse> {
-        let response = self
-            .api()
-            .instance_query_endpoint_delete(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Ok(DeleteResponse {
-            status: response.status,
-            request_id: response.request_id,
-        })
-    }
-
-    // Private endpoint (delegated to library client)
-    pub async fn create_private_endpoint(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::ServicPrivateEndpointePostRequest,
-    ) -> Result<clickhouse_cloud_api::models::InstancePrivateEndpoint> {
-        let response = self
-            .api()
-            .instance_private_endpoint_create(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Private endpoint config (delegated to library client)
-    pub async fn get_service_private_endpoint_config(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::PrivateEndpointConfig> {
-        let response = self
-            .api()
-            .instance_private_endpoint_config_get(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_service_prometheus(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        filtered_metrics: Option<bool>,
-    ) -> Result<String> {
-        let filtered = filtered_metrics.map(|b| b.to_string());
-        self.api()
-            .instance_prometheus_get(org_id, service_id, filtered.as_deref())
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))
-    }
-
     // ClickPipe endpoints (delegated to library client)
     pub async fn list_clickpipes(
         &self,
