+++ crates/clickhousectl/src/cloud/commands.rs
@@ -1,715 +1,7 @@
-use crate::cloud::api_keys::{cleanup_service_query_key, service_query_key_cleanup};
-use crate::cloud::client::{CloudClient, CloudError};
-use crate::cloud::credentials;
-use crate::cloud::output::{ABSENT, or_absent, print_human};
-use crate::cloud::shared::{parse_serde_enum, parse_tags, resolve_org_id};
-use clickhouse_cloud_api::models::{
-    AutoscalingMode, InstancePrivateEndpointsPatch, InstanceServiceQueryApiEndpointsPostRequest,
-    InstanceTagsPatch, IpAccessListEntry, IpAccessListPatch, ServicPrivateEndpointePostRequest,
-    Service, ServiceEndpoint, ServiceEndpointChange, ServiceEndpointChangeProtocol,
-    ServicePasswordPatchRequest, ServicePatchRequest, ServicePatchRequestReleasechannel,
-    ServicePostRequest, ServicePostRequestCompliancetype, ServicePostRequestProfile,
-    ServicePostRequestProvider, ServicePostRequestRegion, ServicePostRequestReleasechannel,
-    ServiceReplicaScalingPatchRequest, ServiceState, ServiceStatePatchRequestCommand,
-};
-use std::io::IsTerminal;
+use crate::cloud::client::CloudClient;
+use crate::cloud::output::{or_absent, print_human};
+use crate::cloud::shared::resolve_org_id;
 use tabled::{Table, Tabled, settings::Style};
-
-/// `host:port` of a service's first endpoint, for list tables.
-///
-/// Renders [`ABSENT`] when the API returned no endpoints, and keeps whichever
-/// half of a partial endpoint it did return (`host` alone, or `-:port`).
-fn first_endpoint(endpoints: Option<&[ServiceEndpoint]>) -> String {
-    endpoints
-        .and_then(|endpoints| endpoints.first())
-        .map(|endpoint| match (endpoint.host.as_deref(), endpoint.port) {
-            (Some(host), Some(port)) => format!("{host}:{port}"),
-            (Some(host), None) => host.to_string(),
-            (None, Some(port)) => format!("{ABSENT}:{port}"),
-            (None, None) => ABSENT.to_string(),
-        })
-        .unwrap_or_else(|| ABSENT.to_string())
-}
-
-/// Resolve a service by name or ID within the given org.
-/// Exactly one of `name` or `id` must be provided.
-async fn resolve_service(
-    client: &CloudClient,
-    org_id: &str,
-    name: Option<&str>,
-    id: Option<&str>,
-) -> Result<Service, Box<dyn std::error::Error>> {
-    match (name, id) {
-        (Some(name), None) => {
-            let services = client.list_services(org_id).await?;
-            let matches: Vec<_> = services
-                .into_iter()
-                .filter(|s| s.name.as_deref() == Some(name))
-                .collect();
-            match matches.len() {
-                0 => Err(format!("no service found with name '{}'", name).into()),
-                1 => Ok(matches.into_iter().next().unwrap()),
-                n => Err(format!(
-                    "found {} services named '{}' — use --id to disambiguate",
-                    n, name
-                )
-                .into()),
-            }
-        }
-        (None, Some(id)) => Ok(client.get_service(org_id, id).await?),
-        (Some(_), Some(_)) => Err("specify either --name or --id, not both".into()),
-        (None, None) => Err("specify --name or --id to identify the service".into()),
-    }
-}
-
-fn parse_ip_access_entries(values: &[String]) -> Option<Vec<IpAccessListEntry>> {
-    (!values.is_empty()).then(|| {
-        values
-            .iter()
-            .map(|value| IpAccessListEntry {
-                source: value.clone(),
-                description: None,
-            })
-            .collect()
-    })
-}
-
-fn parse_ip_access_list_patch(add: &[String], remove: &[String]) -> Option<IpAccessListPatch> {
-    let patch = IpAccessListPatch {
-        add: parse_ip_access_entries(add).unwrap_or_default(),
-        remove: parse_ip_access_entries(remove).unwrap_or_default(),
-    };
-
-    (!patch.add.is_empty() || !patch.remove.is_empty()).then_some(patch)
-}
-
-fn parse_private_endpoint_ids_patch(
-    add: &[String],
-    remove: &[String],
-) -> Option<InstancePrivateEndpointsPatch> {
-    let patch = InstancePrivateEndpointsPatch {
-        add: if add.is_empty() { vec![] } else { add.to_vec() },
-        remove: if remove.is_empty() {
-            vec![]
-        } else {
-            remove.to_vec()
-        },
-    };
-
-    (!patch.add.is_empty() || !patch.remove.is_empty()).then_some(patch)
-}
-
-fn parse_service_endpoint_changes(
-    enable: &[String],
-    disable: &[String],
-) -> Result<Option<Vec<ServiceEndpointChange>>, Box<dyn std::error::Error>> {
-    let mut changes = Vec::new();
-
-    for protocol in enable {
-        changes.push(ServiceEndpointChange {
-            protocol: parse_serde_enum::<ServiceEndpointChangeProtocol>(
-                protocol,
-                "endpoint",
-                &["mysql"],
-            )?,
-            enabled: true,
-        });
-    }
-
-    for protocol in disable {
-        changes.push(ServiceEndpointChange {
-            protocol: parse_serde_enum::<ServiceEndpointChangeProtocol>(
-                protocol,
-                "endpoint",
-                &["mysql"],
-            )?,
-            enabled: false,
-        });
-    }
-
-    Ok((!changes.is_empty()).then_some(changes))
-}
-
-fn parse_instance_tags_patch(
-    add: &[String],
-    remove: &[String],
-) -> Result<Option<InstanceTagsPatch>, Box<dyn std::error::Error>> {
-    let patch = InstanceTagsPatch {
-        add: parse_tags(add)?.unwrap_or_default(),
-        remove: parse_tags(remove)?.unwrap_or_default(),
-    };
-
-    Ok((!patch.add.is_empty() || !patch.remove.is_empty()).then_some(patch))
-}
-
-pub async fn service_list(
-    client: &CloudClient,
-    org_id: Option<&str>,
-    filters: &[String],
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let services = if filters.is_empty() {
-        client.list_services(&org_id).await?
-    } else {
-        client.list_services_filtered(&org_id, filters).await?
-    };
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&services)?);
-    } else {
-        if services.is_empty() {
-            println!("No services found");
-            return Ok(());
-        }
-        #[derive(Tabled)]
-        struct Row {
-            #[tabled(rename = "Name")]
-            name: String,
-            #[tabled(rename = "ID")]
-            id: String,
-            #[tabled(rename = "State")]
-            state: String,
-            #[tabled(rename = "Provider")]
-            provider: String,
-            #[tabled(rename = "Region")]
-            region: String,
-            #[tabled(rename = "Endpoint")]
-            endpoint: String,
-        }
-        let rows: Vec<Row> = services
-            .into_iter()
-            .map(|svc| Row {
-                name: or_absent(svc.name.as_deref()),
-                id: or_absent(svc.id),
-                state: or_absent(svc.state.as_ref()),
-                provider: or_absent(svc.provider.as_ref()),
-                region: or_absent(svc.region.as_ref()),
-                endpoint: first_endpoint(svc.endpoints.as_deref()),
-            })
-            .collect();
-        println!("{}", Table::new(rows).with(Style::markdown()));
-    }
-    Ok(())
-}
-
-pub async fn service_get(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let svc = client.get_service(&org_id, service_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&svc)?);
-    } else {
-        print_human(&svc)?;
-    }
-    Ok(())
-}
-
-/// Options for creating a service
-#[derive(Default)]
-pub struct CreateServiceOptions {
-    pub name: String,
-    pub provider: String,
-    pub region: String,
-    pub min_replica_memory_gb: Option<u32>,
-    pub max_replica_memory_gb: Option<u32>,
-    pub num_replicas: Option<u32>,
-    pub min_replicas: Option<u32>,
-    pub max_replicas: Option<u32>,
-    pub autoscaling_mode: Option<String>,
-    pub idle_scaling: Option<bool>,
-    pub idle_timeout_minutes: Option<u32>,
-    pub ip_allow: Vec<String>,
-    pub backup_id: Option<String>,
-    pub release_channel: Option<String>,
-    pub data_warehouse_id: Option<String>,
-    pub is_readonly: bool,
-    pub encryption_key: Option<String>,
-    pub encryption_role: Option<String>,
-    pub enable_tde: bool,
-    pub compliance_type: Option<String>,
-    pub profile: Option<String>,
-    pub tags: Vec<String>,
-    pub enable_endpoints: Vec<String>,
-    pub disable_endpoints: Vec<String>,
-    pub private_preview_terms_checked: bool,
-    pub enable_core_dumps: Option<bool>,
-    pub org_id: Option<String>,
-}
-
-#[derive(Default)]
-pub struct ServiceUpdateOptions {
-    pub name: Option<String>,
-    pub add_ip_allow: Vec<String>,
-    pub remove_ip_allow: Vec<String>,
-    pub add_private_endpoint_ids: Vec<String>,
-    pub remove_private_endpoint_ids: Vec<String>,
-    pub release_channel: Option<String>,
-    pub enable_endpoints: Vec<String>,
-    pub disable_endpoints: Vec<String>,
-    pub transparent_data_encryption_key_id: Option<String>,
-    pub add_tags: Vec<String>,
-    pub remove_tags: Vec<String>,
-    pub enable_core_dumps: Option<bool>,
-    pub org_id: Option<String>,
-}
-
-#[derive(Default)]
-pub struct ServiceResetPasswordOptions {
-    pub new_password_hash: Option<String>,
-    pub new_double_sha1_hash: Option<String>,
-    pub org_id: Option<String>,
-}
-
-#[derive(Default)]
-pub struct QueryEndpointCreateOptions {
-    pub roles: Vec<String>,
-    pub open_api_keys: Vec<String>,
-    pub allowed_origins: Option<String>,
-    pub org_id: Option<String>,
-}
-
-/// Resolved horizontal-autoscaling fields for a service create/scale request.
-struct HorizontalAutoscaling {
-    autoscaling_mode: Option<AutoscalingMode>,
-    min_replicas: Option<i64>,
-    max_replicas: Option<i64>,
-}
-
-/// Resolve the horizontal-autoscaling fields shared by `service create` and
-/// `service scale`.
-///
-/// The mode is sent only when `--autoscaling-mode` is given explicitly. The
-/// API resolves an omitted mode itself, and a min/max band with the mode
-/// omitted and min == max is accepted as a vertical fixed replica count that
-/// needs no horizontal entitlement — inferring `horizontal` here would change
-/// those semantics.
-///
-/// Rejects `--min-replicas` without `--max-replicas` (and vice versa) with a
-/// clear error before any network call. clap already rejects mixing the
-/// horizontal pair with `--num-replicas`; the memory flags and
-/// `--autoscaling-mode` combine freely with either set because a single
-/// request can switch modes (e.g. `--autoscaling-mode vertical
-/// --num-replicas 3`, or `--autoscaling-mode horizontal` with the equal
-/// memory bounds horizontal requires). Remaining combination rules are the
-/// API's to enforce.
-fn resolve_horizontal_autoscaling(
-    autoscaling_mode: Option<&str>,
-    min_replicas: Option<u32>,
-    max_replicas: Option<u32>,
-) -> Result<HorizontalAutoscaling, Box<dyn std::error::Error>> {
-    if min_replicas.is_some() != max_replicas.is_some() {
-        return Err("--min-replicas and --max-replicas must be specified together".into());
-    }
-    let autoscaling_mode = autoscaling_mode
-        .map(|value| {
-            parse_serde_enum::<AutoscalingMode>(value, "autoscaling_mode", AutoscalingMode::VALUES)
-        })
-        .transpose()?;
-    Ok(HorizontalAutoscaling {
-        autoscaling_mode,
-        min_replicas: min_replicas.map(i64::from),
-        max_replicas: max_replicas.map(i64::from),
-    })
-}
-
-fn build_create_service_request(
-    opts: &CreateServiceOptions,
-) -> Result<ServicePostRequest, Box<dyn std::error::Error>> {
-    let ip_access_list = if opts.ip_allow.is_empty() {
-        vec![IpAccessListEntry {
-            source: "0.0.0.0/0".to_string(),
-            description: Some("Allow all (created by clickhousectl)".to_string()),
-        }]
-    } else {
-        parse_ip_access_entries(&opts.ip_allow).unwrap_or_default()
-    };
-
-    let horizontal = resolve_horizontal_autoscaling(
-        opts.autoscaling_mode.as_deref(),
-        opts.min_replicas,
-        opts.max_replicas,
-    )?;
-
-    Ok(ServicePostRequest {
-        name: opts.name.clone(),
-        provider: parse_serde_enum::<ServicePostRequestProvider>(
-            &opts.provider,
-            "provider",
-            ServicePostRequestProvider::VALUES,
-        )?,
-        region: parse_serde_enum::<ServicePostRequestRegion>(
-            &opts.region,
-            "region",
-            ServicePostRequestRegion::VALUES,
-        )?,
-        ip_access_list,
-        min_replica_memory_gb: opts.min_replica_memory_gb.map(f64::from),
-        max_replica_memory_gb: opts.max_replica_memory_gb.map(f64::from),
-        num_replicas: opts.num_replicas.map(i64::from),
-        idle_scaling: opts.idle_scaling,
-        idle_timeout_minutes: opts.idle_timeout_minutes.map(f64::from),
-        backup_id: opts
-            .backup_id
-            .as_deref()
-            .map(uuid::Uuid::parse_str)
-            .transpose()
-            .map_err(|e| format!("invalid backup_id: {}", e))?,
-        release_channel: match opts.release_channel.as_deref() {
-            Some(value) => Some(parse_serde_enum::<ServicePostRequestReleasechannel>(
-                value,
-                "release_channel",
-                ServicePostRequestReleasechannel::VALUES,
-            )?),
-            None => None,
-        },
-        tags: parse_tags(&opts.tags)?,
-        data_warehouse_id: opts.data_warehouse_id.clone(),
-        is_readonly: if opts.is_readonly { Some(true) } else { None },
-        encryption_key: opts.encryption_key.clone(),
-        encryption_assumed_role_identifier: opts.encryption_role.clone(),
-        has_transparent_data_encryption: if opts.enable_tde { Some(true) } else { None },
-        compliance_type: match opts.compliance_type.as_deref() {
-            Some(value) => Some(parse_serde_enum::<ServicePostRequestCompliancetype>(
-                value,
-                "compliance_type",
-                ServicePostRequestCompliancetype::VALUES,
-            )?),
-            None => None,
-        },
-        profile: match opts.profile.as_deref() {
-            Some(value) => Some(parse_serde_enum::<ServicePostRequestProfile>(
-                value,
-                "profile",
-                ServicePostRequestProfile::VALUES,
-            )?),
-            None => None,
-        },
-        private_preview_terms_checked: if opts.private_preview_terms_checked {
-            Some(true)
-        } else {
-            None
-        },
-        endpoints: parse_service_endpoint_changes(&opts.enable_endpoints, &opts.disable_endpoints)?,
-        enable_core_dumps: opts.enable_core_dumps,
-        // Fields not exposed in CLI
-        autoscaling_mode: horizontal.autoscaling_mode,
-        byoc_id: None,
-        min_replicas: horizontal.min_replicas,
-        max_replicas: horizontal.max_replicas,
-        // Deprecated fields — only exist (and stay None) under the
-        // `deprecated-fields` feature; gated out of the struct otherwise.
-        #[cfg(feature = "deprecated-fields")]
-        max_total_memory_gb: None,
-        #[cfg(feature = "deprecated-fields")]
-        min_total_memory_gb: None,
-        #[cfg(feature = "deprecated-fields")]
-        private_endpoint_ids: None,
-        #[cfg(feature = "deprecated-fields")]
-        tier: None,
-    })
-}
-
-fn build_update_service_request(
-    opts: &ServiceUpdateOptions,
-) -> Result<ServicePatchRequest, Box<dyn std::error::Error>> {
-    Ok(ServicePatchRequest {
-        name: opts.name.clone(),
-        ip_access_list: parse_ip_access_list_patch(&opts.add_ip_allow, &opts.remove_ip_allow),
-        private_endpoint_ids: parse_private_endpoint_ids_patch(
-            &opts.add_private_endpoint_ids,
-            &opts.remove_private_endpoint_ids,
-        ),
-        release_channel: opts
-            .release_channel
-            .as_deref()
-            .map(|value| {
-                parse_serde_enum::<ServicePatchRequestReleasechannel>(
-                    value,
-                    "release_channel",
-                    ServicePatchRequestReleasechannel::VALUES,
-                )
-            })
-            .transpose()?,
-        endpoints: parse_service_endpoint_changes(&opts.enable_endpoints, &opts.disable_endpoints)?,
-        transparent_data_encryption_key_id: opts.transparent_data_encryption_key_id.clone(),
-        tags: parse_instance_tags_patch(&opts.add_tags, &opts.remove_tags)?,
-        enable_core_dumps: opts.enable_core_dumps,
-    })
-}
-
-fn build_service_password_patch_request(
-    opts: &ServiceResetPasswordOptions,
-) -> ServicePasswordPatchRequest {
-    ServicePasswordPatchRequest {
-        new_password_hash: opts.new_password_hash.clone(),
-        new_double_sha1_hash: opts.new_double_sha1_hash.clone(),
-    }
-}
-
-fn build_query_endpoint_create_request(
-    opts: &QueryEndpointCreateOptions,
-) -> InstanceServiceQueryApiEndpointsPostRequest {
-    InstanceServiceQueryApiEndpointsPostRequest {
-        roles: opts.roles.clone(),
-        open_api_keys: opts.open_api_keys.clone(),
-        allowed_origins: opts.allowed_origins.clone().unwrap_or_default(),
-    }
-}
-
-/// The post-create hint showing how to query the new service.
-///
-/// The hint is only useful with a real service id: an absent id would render a
-/// command line the user cannot run, so the hint is dropped rather than printed
-/// with a placeholder id in it.
-fn service_query_hint(service_id: Option<uuid::Uuid>) -> Option<String> {
-    service_id.map(|id| {
-        format!(
-            "Run SQL with: clickhousectl cloud service query --id {} --query \"SELECT 1\"\n\
-             (the Query API endpoint is provisioned automatically on first use)",
-            id
-        )
-    })
-}
-
-/// The post-create credentials block, or the warning that replaces it.
-///
-/// The generated password is returned once, so a placeholder in its place would
-/// be read as the credential itself. An absent password therefore gets the
-/// omission plus the command that mints a usable one; the create succeeded and
-/// the password is recoverable, so this is a warning rather than an error. An
-/// empty string is a password the API sent.
-fn service_credentials_block(password: Option<&str>, service_id: Option<uuid::Uuid>) -> String {
-    match (password, service_id) {
-        (Some(password), _) => format!(
-            "Credentials (save these, password shown only once):\n  Username: default\n  \
-             Password: {}",
-            password
-        ),
-        (None, Some(id)) => format!(
-            "WARNING: the API response omitted the one-time password, so it cannot be shown.\n\
-             The service was created; reset the password to get a usable credential:\n  \
-             clickhousectl cloud service reset-password {}",
-            id
-        ),
-        (None, None) => "WARNING: the API response omitted the one-time password, so it cannot be \
-                         shown.\nThe service was created; once you have its id, reset the password \
-                         with `clickhousectl cloud service reset-password <service-id>` to get a \
-                         usable credential."
-            .to_string(),
-    }
-}
-
-pub async fn service_create(
-    client: &CloudClient,
-    opts: CreateServiceOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Validate input before any network call so typos like --provider awss
-    // fail locally instead of on the /organizations lookup.
-    let request = build_create_service_request(&opts)?;
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-
-    let response = client.create_service(&org_id, &request).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&response)?);
-    } else {
-        let service_id = response.service.as_ref().and_then(|svc| svc.id);
-        println!("Service created successfully!");
-        println!();
-        if let Some(service) = &response.service {
-            print_human(service)?;
-        }
-        println!();
-        println!(
-            "{}",
-            service_credentials_block(response.password.as_deref(), service_id)
-        );
-        if let Some(hint) = service_query_hint(service_id) {
-            println!();
-            println!("{}", hint);
-        }
-    }
-    Ok(())
-}
-
-/// Classifies one poll of a service's state while waiting for a stop to land.
-///
-/// Returns `true` once the service has stopped. An absent state cannot be
-/// classified and the loop has no other exit, so treating it as "not stopped
-/// yet" would poll forever: fail instead of waiting on a state the API is not
-/// reporting.
-fn classify_stop_poll_state(
-    state: Option<&ServiceState>,
-) -> Result<bool, Box<dyn std::error::Error>> {
-    let state = state
-        .ok_or(
-            "the API response omitted the service state while waiting for the service to stop, \
-             so the stop cannot be confirmed",
-        )?
-        .to_string();
-    if matches!(state.as_str(), "stopped" | "idle") {
-        return Ok(true);
-    }
-    if matches!(state.as_str(), "terminated" | "failed" | "deleted") {
-        return Err(format!(
-            "service entered unexpected state '{}' while waiting for stop",
-            state
-        )
-        .into());
-    }
-    Ok(false)
-}
-
-#[derive(Default)]
-struct StopPollProgress {
-    previous_state: Option<String>,
-}
-
-impl StopPollProgress {
-    fn render(&mut self, state: &str, verbose: bool) -> Option<String> {
-        let changed = self.previous_state.as_deref() != Some(state);
-        self.previous_state = Some(state.to_string());
-        (verbose || changed).then(|| format!("  state: {state}"))
-    }
-}
-
-/// Replace the API's running-service conflict with the CLI remedy.
-///
-/// Keep other conflicts intact: the API can reject deletion for reasons that
-/// `--force` cannot fix, and a forced deletion must not suggest the flag the
-/// user already passed.
-fn service_delete_error(error: CloudError, force: bool, service_id: &str) -> CloudError {
-    if !force
-        && error.message.starts_with("CONFLICT:")
-        && error.message.contains("Current state: 'running'")
-    {
-        CloudError::new(format!(
-            "service is running and cannot be deleted. Use --force to stop it first, or \
-             `clickhousectl cloud service stop {service_id}`."
-        ))
-    } else {
-        error
-    }
-}
-
-pub async fn service_delete(
-    client: &CloudClient,
-    service_id: &str,
-    force: bool,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-    let (query_key_id, retain_query_key) = service_query_key_cleanup(&org_id, service_id)?;
-
-    if force {
-        let svc = client.get_service_if_exists(&org_id, service_id).await?;
-        // An absent state matches nothing: skip the stop and let the delete
-        // call decide, rather than guessing the service is running.
-        let state = svc
-            .as_ref()
-            .map(|service| or_absent(service.state.as_ref()))
-            .unwrap_or_default();
-        if matches!(state.as_str(), "running" | "idle" | "starting") {
-            eprintln!("Stopping service {} before deletion...", service_id);
-            client
-                .change_service_state(&org_id, service_id, ServiceStatePatchRequestCommand::Stop)
-                .await?;
-
-            // Poll until the service is stopped
-            let verbose_polling =
-                std::io::stderr().is_terminal() && !json && std::env::var_os("CI").is_none();
-            let mut progress = StopPollProgress::default();
-            loop {
-                tokio::time::sleep(std::time::Duration::from_secs(5)).await;
-                let svc = client.get_service(&org_id, service_id).await?;
-                let state = or_absent(svc.state.as_ref());
-                if let Some(line) = progress.render(&state, verbose_polling) {
-                    eprintln!("{line}");
-                }
-                if classify_stop_poll_state(svc.state.as_ref())? {
-                    break;
-                }
-            }
-        }
-    }
-
-    let response = client
-        .delete_service_if_exists(&org_id, service_id)
-        .await
-        .map_err(|error| service_delete_error(error, force, service_id))?;
-    // Delete the key only after the service is gone. If cleanup fails, retain
-    // its exact IDs locally so repeating service delete can retry safely.
-    cleanup_service_query_key(client, &org_id, service_id, query_key_id.as_deref()).await?;
-    if !retain_query_key {
-        credentials::remove_service_query_key(service_id)?;
-    }
-    if json {
-        println!("{}", serde_json::to_string_pretty(&response)?);
-    } else if response.is_none() {
-        println!("Service {} is already absent", service_id);
-    } else {
-        println!("Service {} deletion initiated", service_id);
-    }
-    Ok(())
-}
-
-pub async fn service_start(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let svc = client
-        .change_service_state(&org_id, service_id, ServiceStatePatchRequestCommand::Start)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&svc)?);
-    } else {
-        println!(
-            "Service {} starting (state: {})",
-            or_absent(svc.name.as_deref()),
-            or_absent(svc.state.as_ref())
-        );
-    }
-    Ok(())
-}
-
-pub async fn service_stop(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let svc = client
-        .change_service_state(&org_id, service_id, ServiceStatePatchRequestCommand::Stop)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&svc)?);
-    } else {
-        println!(
-            "Service {} stopping (state: {})",
-            or_absent(svc.name.as_deref()),
-            or_absent(svc.state.as_ref())
-        );
-    }
-    Ok(())
-}
 
 pub async fn clickpipe_list(
     client: &CloudClient,
@@ -1690,1342 +982,9 @@ pub async fn clickpipe_create_bigquery(
     Ok(())
 }
 
-pub async fn service_update(
-    client: &CloudClient,
-    service_id: &str,
-    opts: ServiceUpdateOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Validate input before any network call so typos like --release-channel turbo
-    // fail locally instead of on the /organizations lookup.
-    let request = build_update_service_request(&opts)?;
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-
-    let svc = client.update_service(&org_id, service_id, &request).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&svc)?);
-    } else {
-        println!("Service {} updated", or_absent(svc.name.as_deref()));
-        println!("  ID: {}", or_absent(svc.id));
-        println!("  State: {}", or_absent(svc.state.as_ref()));
-    }
-    Ok(())
-}
-
-#[derive(Default)]
-pub struct ServiceScaleOptions {
-    pub min_replica_memory_gb: Option<u32>,
-    pub max_replica_memory_gb: Option<u32>,
-    pub num_replicas: Option<u32>,
-    pub min_replicas: Option<u32>,
-    pub max_replicas: Option<u32>,
-    pub autoscaling_mode: Option<String>,
-    pub idle_scaling: Option<bool>,
-    pub idle_timeout_minutes: Option<u32>,
-    pub org_id: Option<String>,
-}
-
-fn build_service_scale_request(
-    opts: &ServiceScaleOptions,
-) -> Result<ServiceReplicaScalingPatchRequest, Box<dyn std::error::Error>> {
-    let horizontal = resolve_horizontal_autoscaling(
-        opts.autoscaling_mode.as_deref(),
-        opts.min_replicas,
-        opts.max_replicas,
-    )?;
-
-    Ok(ServiceReplicaScalingPatchRequest {
-        autoscaling_mode: horizontal.autoscaling_mode,
-        min_replica_memory_gb: opts.min_replica_memory_gb.map(f64::from),
-        max_replica_memory_gb: opts.max_replica_memory_gb.map(f64::from),
-        min_replicas: horizontal.min_replicas,
-        max_replicas: horizontal.max_replicas,
-        num_replicas: opts.num_replicas.map(i64::from),
-        idle_scaling: opts.idle_scaling,
-        idle_timeout_minutes: opts.idle_timeout_minutes.map(f64::from),
-    })
-}
-
-pub async fn service_scale(
-    client: &CloudClient,
-    service_id: &str,
-    opts: ServiceScaleOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let request = build_service_scale_request(&opts)?;
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-
-    let svc = client
-        .update_replica_scaling(&org_id, service_id, &request)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&svc)?);
-    } else {
-        println!("Service {} scaling updated", or_absent(svc.name.as_deref()));
-        println!(
-            "  Autoscaling Mode: {}",
-            or_absent(svc.autoscaling_mode.as_ref())
-        );
-        match svc.autoscaling_mode {
-            Some(AutoscalingMode::Horizontal) => {
-                println!("  Min Replicas: {}", or_absent(svc.min_replicas));
-                println!("  Max Replicas: {}", or_absent(svc.max_replicas));
-                println!("  Memory/Replica: {} GB", or_absent(svc.replica_memory_gb));
-            }
-            Some(AutoscalingMode::Vertical) => {
-                println!(
-                    "  Min Memory/Replica: {} GB",
-                    or_absent(svc.min_replica_memory_gb)
-                );
-                println!(
-                    "  Max Memory/Replica: {} GB",
-                    or_absent(svc.max_replica_memory_gb)
-                );
-                println!("  Replicas: {}", or_absent(svc.num_replicas));
-            }
-            // A mode this CLI version doesn't know, or one the API did not
-            // return; don't guess which fields apply.
-            _ => {}
-        }
-    }
-    Ok(())
-}
-
-pub async fn service_reset_password(
-    client: &CloudClient,
-    service_id: &str,
-    opts: ServiceResetPasswordOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-    let request = build_service_password_patch_request(&opts);
-    let resp = client.reset_password(&org_id, service_id, &request).await?;
-
-    // Resolve before either output branch, so --json cannot report success
-    // over a response that dropped the one-time generated password.
-    let outcome =
-        resolve_reset_password_outcome(generation_requested(&request), resp.password.as_deref())?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&resp)?);
-    } else {
-        println!("Password reset for service {}", service_id);
-        match outcome {
-            ResetPasswordOutcome::Generated(password) => {
-                println!("  New password: {}", password)
-            }
-            ResetPasswordOutcome::HashUpdated => {
-                println!("  Password hash updated; no plaintext password returned")
-            }
-        }
-    }
-    Ok(())
-}
-
-/// What a `service reset-password` response says about the new credential.
-#[derive(Debug)]
-enum ResetPasswordOutcome<'a> {
-    /// The API generated the password; it is returned once and never again.
-    Generated(&'a str),
-    /// The caller supplied a hash, so the API generates no plaintext password.
-    HashUpdated,
-}
-
-/// Whether the PATCH body asks the API to generate a password.
-///
-/// `newPasswordHash` alone decides it: the spec says `newDoubleSha1Hash` "will
-/// be ignored and the generated password will be used" when `newPasswordHash`
-/// is absent, and that the response carries a password "only if there was no
-/// 'newPasswordHash' in the request". So a double-SHA1-only body is still a
-/// generation request, and treating it as a hash update would discard the
-/// generated password the API rotated to.
-fn generation_requested(request: &ServicePasswordPatchRequest) -> bool {
-    request.new_password_hash.is_none()
-}
-
-/// Resolves what to report from the request mode and the response.
-///
-/// The mode is read from the request rather than inferred from what came back:
-/// only a hash request explains a response without a password, and an absent
-/// one on a generation request means the new credential is lost.
-fn resolve_reset_password_outcome(
-    generation_requested: bool,
-    password: Option<&str>,
-) -> Result<ResetPasswordOutcome<'_>, Box<dyn std::error::Error>> {
-    if !generation_requested {
-        return Ok(ResetPasswordOutcome::HashUpdated);
-    }
-    match password {
-        Some(password) => Ok(ResetPasswordOutcome::Generated(password)),
-        None => Err(
-            "the API response omitted the generated password, so it cannot be shown: the \
-                     service password may already have been rotated — run the reset again to get \
-                     a password you can use"
-                .into(),
-        ),
-    }
-}
-
-pub async fn query_endpoint_get(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let ep = client.get_query_endpoint(&org_id, service_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&ep)?);
-    } else {
-        print_human(&ep)?;
-    }
-    Ok(())
-}
-
-pub async fn query_endpoint_create(
-    client: &CloudClient,
-    service_id: &str,
-    opts: QueryEndpointCreateOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-    let request = build_query_endpoint_create_request(&opts);
-
-    let ep = client
-        .create_query_endpoint(&org_id, service_id, &request)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&ep)?);
-    } else {
-        println!("Query endpoint created for service {}", service_id);
-        println!("  ID: {}", or_absent(ep.id.as_deref()));
-        println!(
-            "  Roles: {}",
-            or_absent(ep.roles.as_ref().map(|roles| roles.join(", ")))
-        );
-    }
-    Ok(())
-}
-
-pub async fn query_endpoint_delete(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let response = client.delete_query_endpoint(&org_id, service_id).await?;
-    if json {
-        println!("{}", serde_json::to_string_pretty(&response)?);
-    } else {
-        println!("Query endpoint deleted for service {}", service_id);
-    }
-    Ok(())
-}
-
-pub struct ServiceQueryOptions {
-    pub name: Option<String>,
-    pub id: Option<String>,
-    pub query: Option<String>,
-    pub queries_file: Option<String>,
-    pub database: Option<String>,
-    pub format: Option<String>,
-    pub json: bool,
-    pub org_id: Option<String>,
-    pub no_auto_enable: bool,
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn run_basic_service_query(
-    client: &CloudClient,
-    service_id: &str,
-    key_id: &str,
-    key_secret: &str,
-    sql: &str,
-    database: Option<&str>,
-    format: &str,
-    service_name: &str,
-) -> Result<reqwest::Response, clickhouse_cloud_api::Error> {
-    let run = |wake: bool| {
-        client
-            .api()
-            .run_query(service_id, key_id, key_secret, sql, database, format, wake)
-    };
-    match run(false).await {
-        Err(clickhouse_cloud_api::Error::ServiceIdle) => {
-            eprint_waking_service(service_name);
-            run(true).await
-        }
-        other => other,
-    }
-}
-
-fn query_requires_provisioning(error: &clickhouse_cloud_api::Error) -> bool {
-    matches!(
-        error,
-        clickhouse_cloud_api::Error::Api {
-            status: 401 | 403 | 404,
-            ..
-        }
-    )
-}
-
-pub async fn service_query(
-    client: &CloudClient,
-    opts: ServiceQueryOptions,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let sql = read_query_sql(opts.query.as_deref(), opts.queries_file.as_deref())?;
-
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-    let service =
-        resolve_service(client, &org_id, opts.name.as_deref(), opts.id.as_deref()).await?;
-    // The whole query path is keyed on the service id: prefer the one the API
-    // echoed, and fall back to the one the user passed if the response omitted
-    // it.
-    let service_id = match service.id {
-        Some(id) => id.to_string(),
-        None => opts
-            .id
-            .clone()
-            .ok_or("the API response is missing the service id")?,
-    };
-    let service_name = or_absent(service.name.as_deref());
-
-    // An explicit format always wins over agent-triggered JSON mode. Clap
-    // rejects an explicit --json together with --format.
-    let format = opts.format.unwrap_or_else(|| {
-        if opts.json {
-            "JSONEachRow".to_string()
-        } else {
-            default_query_format()
-        }
-    });
-
-    let response = if client.is_bearer_auth() {
-        // OAuth: the Query API authenticates the user's bearer token
-        // directly, SQL-console style — the query runs as the user's own
-        // cloud identity, with no per-service Query API key and no
-        // query-endpoint configuration needed on the service.
-        // `--no-auto-enable` is a no-op here since nothing is ever
-        // provisioned.
-        let run = |wake: bool| {
-            client.api().run_query_bearer(
-                &service_id,
-                &sql,
-                opts.database.as_deref(),
-                &format,
-                wake,
-            )
-        };
-        let result = match run(false).await {
-            Err(clickhouse_cloud_api::Error::ServiceIdle) => {
-                eprint_waking_service(&service_name);
-                run(true).await
-            }
-            other => other,
-        };
-        result.map_err(|e| convert_query_error(client, e, &service_name))?
-    } else {
-        let result = if let Some(key) = credentials::get_service_query_key(&service_id) {
-            run_basic_service_query(
-                client,
-                &service_id,
-                &key.key_id,
-                &key.key_secret,
-                &sql,
-                opts.database.as_deref(),
-                &format,
-                &service_name,
-            )
-            .await
-        } else {
-            let (key_id, key_secret) = client
-                .basic_auth_credentials()
-                .ok_or("API key credentials are unavailable")?;
-            match run_basic_service_query(
-                client,
-                &service_id,
-                key_id,
-                key_secret,
-                &sql,
-                opts.database.as_deref(),
-                &format,
-                &service_name,
-            )
-            .await
-            {
-                Err(error) if query_requires_provisioning(&error) => {
-                    if opts.no_auto_enable {
-                        return Err(format!(
-                            "the authenticated API key cannot use the Query API endpoint for service {service_id}, and --no-auto-enable prevents provisioning"
-                        )
-                        .into());
-                    }
-                    eprintln!(
-                        "Provisioning Query API endpoint + key for service '{}'...",
-                        service_name
-                    );
-                    let key = crate::cloud::service_query::ensure_service_query_setup(
-                        client,
-                        &org_id,
-                        &service_id,
-                        &service_name,
-                    )
-                    .await?;
-                    run_basic_service_query(
-                        client,
-                        &service_id,
-                        &key.key_id,
-                        &key.key_secret,
-                        &sql,
-                        opts.database.as_deref(),
-                        &format,
-                        &service_name,
-                    )
-                    .await
-                }
-                other => other,
-            }
-        };
-        result.map_err(|e| convert_query_error(client, e, &service_name))?
-    };
-
-    use futures_util::StreamExt;
-    use std::io::Write as _;
-    let mut stream = response.bytes_stream();
-    let stdout = std::io::stdout();
-    let mut handle = stdout.lock();
-    let mut byte_count = 0;
-    let mut last_byte = None;
-    while let Some(chunk) = stream.next().await {
-        let bytes = chunk.map_err(|e| -> Box<dyn std::error::Error> {
-            format!("Failed to read query response: {e}").into()
-        })?;
-        handle.write_all(&bytes)?;
-        byte_count += bytes.len();
-        if let Some(last) = bytes.last() {
-            last_byte = Some(*last);
-        }
-    }
-    match query_output_completion(&format, byte_count, last_byte) {
-        QueryOutputCompletion::None => {}
-        QueryOutputCompletion::Newline => handle.write_all(b"\n")?,
-        QueryOutputCompletion::Acknowledge => eprintln!("OK"),
-    }
-    handle.flush()?;
-    Ok(())
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum QueryOutputCompletion {
-    None,
-    Newline,
-    Acknowledge,
-}
-
-fn query_output_completion(
-    format: &str,
-    byte_count: usize,
-    last_byte: Option<u8>,
-) -> QueryOutputCompletion {
-    if byte_count == 0 {
-        // Keep an empty SELECT or DDL response out of the row stream while
-        // still making successful no-output statements visible to the user.
-        return QueryOutputCompletion::Acknowledge;
-    }
-    if query_format_uses_text_lines(format) && last_byte != Some(b'\n') {
-        QueryOutputCompletion::Newline
-    } else {
-        QueryOutputCompletion::None
-    }
-}
-
-fn query_format_uses_text_lines(format: &str) -> bool {
-    // Preserve unknown and customizable formats byte-for-byte rather than
-    // guessing that anything not known to be binary is safe to normalize.
-    matches!(
-        format.to_ascii_lowercase().as_str(),
-        "jsoneachrow" | "prettycompact" | "tabseparated"
-    )
-}
-
-/// Stderr notice shown when the query host reports the service is idled and
-/// the CLI resends the query with the wake confirmation.
-fn eprint_waking_service(service_name: &str) {
-    eprintln!("Service '{service_name}' is idle; waking it (this may take a minute)...");
-}
-
-/// Map Query API errors to user-facing messages: a stopped service gets a
-/// hint to start it (the query host never wakes a stopped service), the
-/// rest go through the standard cloud error conversion.
-fn convert_query_error(
-    client: &CloudClient,
-    err: clickhouse_cloud_api::Error,
-    service_name: &str,
-) -> Box<dyn std::error::Error> {
-    match err {
-        clickhouse_cloud_api::Error::ServiceStopped => format!(
-            "service '{service_name}' is stopped; start it with `clickhousectl cloud service start` and retry"
-        )
-        .into(),
-        other => client.convert_error(other).into(),
-    }
-}
-
-fn read_query_sql(
-    inline: Option<&str>,
-    queries_file: Option<&str>,
-) -> Result<String, Box<dyn std::error::Error>> {
-    use std::io::Read as _;
-
-    if let Some(q) = inline {
-        let trimmed = q.trim();
-        if trimmed.is_empty() {
-            return Err("--query was empty".into());
-        }
-        return Ok(q.to_string());
-    }
-
-    if let Some(path) = queries_file {
-        let mut content = String::new();
-        if path == "-" {
-            std::io::stdin().read_to_string(&mut content)?;
-        } else {
-            content = std::fs::read_to_string(path)?;
-        }
-        if content.trim().is_empty() {
-            return Err("queries file was empty".into());
-        }
-        return Ok(content);
-    }
-
-    if std::io::stdin().is_terminal() {
-        return Err("no SQL provided. Pass --query, --queries-file, or pipe SQL on stdin.".into());
-    }
-
-    let mut content = String::new();
-    std::io::stdin().read_to_string(&mut content)?;
-    if content.trim().is_empty() {
-        return Err("no SQL received on stdin".into());
-    }
-    Ok(content)
-}
-
-fn default_query_format() -> String {
-    if std::io::stdout().is_terminal() {
-        "PrettyCompact".to_string()
-    } else {
-        "TabSeparated".to_string()
-    }
-}
-
-pub async fn private_endpoint_create(
-    client: &CloudClient,
-    service_id: &str,
-    endpoint_id: &str,
-    description: Option<&str>,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let request = ServicPrivateEndpointePostRequest {
-        id: endpoint_id.to_string(),
-        description: description.map(String::from).unwrap_or_default(),
-    };
-
-    let ep = client
-        .create_private_endpoint(&org_id, service_id, &request)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&ep)?);
-    } else {
-        println!("Private endpoint created for service {}", service_id);
-        println!("  Endpoint ID: {}", or_absent(ep.id.as_deref()));
-        println!("  Description: {}", or_absent(ep.description.as_deref()));
-    }
-    Ok(())
-}
-
-pub async fn private_endpoint_get_config(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-    let config = client
-        .get_service_private_endpoint_config(&org_id, service_id)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&config)?);
-    } else {
-        print_human(&config)?;
-    }
-    Ok(())
-}
-
-pub async fn service_prometheus(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    filtered_metrics: Option<bool>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-    let prom = client
-        .get_service_prometheus(&org_id, service_id, filtered_metrics)
-        .await?;
-    println!("{}", prom);
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn first_endpoint_keeps_the_present_half_of_a_partial_endpoint() {
-        let endpoint = |host: Option<&str>, port: Option<i64>| ServiceEndpoint {
-            host: host.map(str::to_string),
-            port,
-            ..Default::default()
-        };
-
-        assert_eq!(first_endpoint(None), ABSENT);
-        assert_eq!(first_endpoint(Some(&[])), ABSENT);
-        assert_eq!(
-            first_endpoint(Some(&[endpoint(Some("host"), Some(9440))])),
-            "host:9440"
-        );
-        assert_eq!(
-            first_endpoint(Some(&[endpoint(Some("host"), None)])),
-            "host"
-        );
-        assert_eq!(
-            first_endpoint(Some(&[endpoint(None, Some(9440))])),
-            format!("{ABSENT}:9440")
-        );
-        assert_eq!(first_endpoint(Some(&[endpoint(None, None)])), ABSENT);
-    }
-
-    #[test]
-    fn service_query_hint_is_dropped_when_the_id_is_absent() {
-        assert_eq!(service_query_hint(None), None);
-
-        let id = uuid::Uuid::parse_str("a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6").unwrap();
-        let hint = service_query_hint(Some(id)).unwrap();
-        assert!(
-            hint.contains("--id a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6"),
-            "hint should name the service: {hint}"
-        );
-        assert!(hint.contains("provisioned automatically on first use"));
-    }
-
-    #[test]
-    fn service_credentials_block_shows_the_password_the_api_sent() {
-        let id = uuid::Uuid::parse_str("a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6").unwrap();
-        assert_eq!(
-            service_credentials_block(Some("s3cret"), Some(id)),
-            "Credentials (save these, password shown only once):\n  Username: default\n  \
-             Password: s3cret"
-        );
-    }
-
-    #[test]
-    fn service_credentials_block_treats_an_empty_password_as_sent() {
-        assert_eq!(
-            service_credentials_block(Some(""), None),
-            "Credentials (save these, password shown only once):\n  Username: default\n  \
-             Password: "
-        );
-    }
-
-    #[test]
-    fn service_credentials_block_warns_with_the_reset_command_when_the_password_is_absent() {
-        let id = uuid::Uuid::parse_str("a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6").unwrap();
-        let block = service_credentials_block(None, Some(id));
-        assert!(
-            !block.contains(&format!("Password: {ABSENT}")),
-            "an absent password must not render a placeholder credential: {block}"
-        );
-        assert!(block.starts_with("WARNING: the API response omitted the one-time password"));
-        assert!(
-            block.contains(
-                "clickhousectl cloud service reset-password a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6"
-            ),
-            "the warning should name the exact recovery command: {block}"
-        );
-    }
-
-    #[test]
-    fn service_credentials_block_warns_generically_when_the_service_id_is_absent() {
-        let block = service_credentials_block(None, None);
-        assert!(block.starts_with("WARNING: the API response omitted the one-time password"));
-        assert!(
-            block.contains("clickhousectl cloud service reset-password <service-id>"),
-            "without an id the warning should stay generic: {block}"
-        );
-    }
-
-    #[test]
-    fn classify_stop_poll_state_fails_on_an_absent_state() {
-        let err = classify_stop_poll_state(None).unwrap_err();
-        assert_eq!(
-            err.to_string(),
-            "the API response omitted the service state while waiting for the service to stop, \
-             so the stop cannot be confirmed"
-        );
-    }
-
-    #[test]
-    fn classify_stop_poll_state_separates_stopped_waiting_and_failed() {
-        assert!(classify_stop_poll_state(Some(&ServiceState::Stopped)).unwrap());
-        assert!(classify_stop_poll_state(Some(&ServiceState::Idle)).unwrap());
-        assert!(!classify_stop_poll_state(Some(&ServiceState::Stopping)).unwrap());
-        assert!(!classify_stop_poll_state(Some(&ServiceState::Running)).unwrap());
-        // An unrecognized state keeps the loop polling rather than failing.
-        assert!(
-            !classify_stop_poll_state(Some(&ServiceState::Unknown("hibernating".into()))).unwrap()
-        );
-
-        let err = classify_stop_poll_state(Some(&ServiceState::Failed)).unwrap_err();
-        assert_eq!(
-            err.to_string(),
-            "service entered unexpected state 'failed' while waiting for stop"
-        );
-        // "deleted" is not a typed variant, so it arrives through the catch-all.
-        let err =
-            classify_stop_poll_state(Some(&ServiceState::Unknown("deleted".into()))).unwrap_err();
-        assert_eq!(
-            err.to_string(),
-            "service entered unexpected state 'deleted' while waiting for stop"
-        );
-    }
-
-    #[test]
-    fn stop_poll_progress_collapses_repeats_but_keeps_transitions() {
-        let mut progress = StopPollProgress::default();
-        assert_eq!(
-            progress.render("stopping", false).as_deref(),
-            Some("  state: stopping")
-        );
-        assert_eq!(progress.render("stopping", false), None);
-        assert_eq!(
-            progress.render("running", false).as_deref(),
-            Some("  state: running")
-        );
-        assert_eq!(
-            progress.render("stopped", false).as_deref(),
-            Some("  state: stopped")
-        );
-    }
-
-    #[test]
-    fn stop_poll_progress_keeps_repeats_in_verbose_mode() {
-        let mut progress = StopPollProgress::default();
-        assert!(progress.render("stopping", true).is_some());
-        assert!(progress.render("stopping", true).is_some());
-    }
-
-    #[test]
-    fn query_output_completion_acknowledges_empty_responses() {
-        assert_eq!(
-            query_output_completion("TabSeparated", 0, None),
-            QueryOutputCompletion::Acknowledge
-        );
-        assert_eq!(
-            query_output_completion("RowBinary", 0, None),
-            QueryOutputCompletion::Acknowledge
-        );
-    }
-
-    #[test]
-    fn query_provisions_only_for_endpoint_or_auth_rejections() {
-        for status in [401, 403, 404] {
-            assert!(query_requires_provisioning(
-                &clickhouse_cloud_api::Error::Api {
-                    status,
-                    message: "rejected".into(),
-                }
-            ));
-        }
-        for status in [400, 408, 429, 500] {
-            assert!(!query_requires_provisioning(
-                &clickhouse_cloud_api::Error::Api {
-                    status,
-                    message: "query failed".into(),
-                }
-            ));
-        }
-        assert!(!query_requires_provisioning(
-            &clickhouse_cloud_api::Error::ServiceStopped
-        ));
-    }
-
-    #[test]
-    fn query_output_completion_adds_only_a_missing_text_newline() {
-        for format in ["TabSeparated", "PrettyCompact", "JSONEachRow"] {
-            assert_eq!(
-                query_output_completion(format, 2, Some(b'K')),
-                QueryOutputCompletion::Newline
-            );
-        }
-        assert_eq!(
-            query_output_completion("JSONEachRow", 2, Some(b'\n')),
-            QueryOutputCompletion::None
-        );
-    }
-
-    #[test]
-    fn query_output_completion_preserves_exact_and_binary_bodies() {
-        for format in [
-            "Template",
-            "Buffers",
-            "BSONEachRow",
-            "RowBinary",
-            "Native",
-            "Parquet",
-            "ArrowStream",
-            "MsgPack",
-        ] {
-            assert_eq!(
-                query_output_completion(format, 3, Some(0)),
-                QueryOutputCompletion::None,
-                "{format} output must stay byte-for-byte intact"
-            );
-        }
-    }
-
-    #[test]
-    fn service_delete_error_suggests_force_for_a_running_service() {
-        let error = CloudError::new(
-            "CONFLICT: Only instance in one of the following states can be terminated. \
-             Current state: 'running'",
-        );
-
-        let error = service_delete_error(error, false, "svc-1");
-
-        assert_eq!(
-            error.message,
-            "service is running and cannot be deleted. Use --force to stop it first, or \
-             `clickhousectl cloud service stop svc-1`."
-        );
-    }
-
-    #[test]
-    fn service_delete_error_preserves_unrelated_and_forced_failures() {
-        let unrelated = CloudError::new("CONFLICT: service has dependent resources");
-        assert_eq!(
-            service_delete_error(unrelated, false, "svc-1").message,
-            "CONFLICT: service has dependent resources"
-        );
-
-        let forced = CloudError::new("CONFLICT: Current state: 'running'");
-        assert_eq!(
-            service_delete_error(forced, true, "svc-1").message,
-            "CONFLICT: Current state: 'running'"
-        );
-    }
-
-    #[test]
-    fn generation_is_requested_unless_a_password_hash_is_sent() {
-        let request = |new_password_hash: Option<&str>, new_double_sha1_hash: Option<&str>| {
-            ServicePasswordPatchRequest {
-                new_password_hash: new_password_hash.map(str::to_string),
-                new_double_sha1_hash: new_double_sha1_hash.map(str::to_string),
-            }
-        };
-        assert!(generation_requested(&request(None, None)));
-        // The API ignores `newDoubleSha1Hash` without `newPasswordHash` and
-        // generates a password anyway, so this is still a generation request:
-        // reporting "hash updated" would discard the new credential.
-        assert!(generation_requested(&request(None, Some("sha1"))));
-        assert!(!generation_requested(&request(Some("sha256"), None)));
-        assert!(!generation_requested(&request(
-            Some("sha256"),
-            Some("sha1")
-        )));
-    }
-
-    #[test]
-    fn resolve_reset_password_outcome_returns_the_generated_password() {
-        match resolve_reset_password_outcome(true, Some("s3cret")).unwrap() {
-            ResetPasswordOutcome::Generated(password) => assert_eq!(password, "s3cret"),
-            ResetPasswordOutcome::HashUpdated => panic!("expected the generated password"),
-        }
-
-        // An empty string is a password the API did send, so it is not absent.
-        assert!(matches!(
-            resolve_reset_password_outcome(true, Some("")).unwrap(),
-            ResetPasswordOutcome::Generated("")
-        ));
-    }
-
-    #[test]
-    fn resolve_reset_password_outcome_reports_hash_updated_regardless_of_response() {
-        assert!(matches!(
-            resolve_reset_password_outcome(false, None).unwrap(),
-            ResetPasswordOutcome::HashUpdated
-        ));
-        // The mode comes from the request, so an echoed password does not
-        // turn a hash reset into a generated one.
-        assert!(matches!(
-            resolve_reset_password_outcome(false, Some("s3cret")).unwrap(),
-            ResetPasswordOutcome::HashUpdated
-        ));
-    }
-
-    #[test]
-    fn resolve_reset_password_outcome_fails_when_the_generated_password_is_absent() {
-        let err = resolve_reset_password_outcome(true, None)
-            .unwrap_err()
-            .to_string();
-        assert!(
-            err.contains("omitted the generated password"),
-            "unexpected error: {err}"
-        );
-        assert!(
-            err.contains("may already have been rotated") && err.contains("run the reset again"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn build_create_service_request_supports_ga_optional_fields() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replica_memory_gb: Some(24),
-            max_replica_memory_gb: Some(48),
-            num_replicas: Some(3),
-            min_replicas: None,
-            max_replicas: None,
-            autoscaling_mode: None,
-            idle_scaling: Some(true),
-            idle_timeout_minutes: Some(10),
-            ip_allow: vec!["10.0.0.0/8".to_string()],
-            backup_id: Some("a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6".to_string()),
-            release_channel: Some("fast".to_string()),
-            data_warehouse_id: Some("dw-1".to_string()),
-            is_readonly: true,
-            encryption_key: Some("key-1".to_string()),
-            encryption_role: Some("role-1".to_string()),
-            enable_tde: true,
-            compliance_type: Some("hipaa".to_string()),
-            profile: Some("v1-default".to_string()),
-            tags: vec!["env=prod".to_string()],
-            enable_endpoints: vec!["mysql".to_string()],
-            disable_endpoints: vec![],
-            private_preview_terms_checked: true,
-            enable_core_dumps: Some(true),
-            org_id: None,
-        };
-
-        let request = build_create_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["tags"][0]["key"], "env");
-        assert_eq!(json["endpoints"][0]["protocol"], "mysql");
-        assert_eq!(json["privatePreviewTermsChecked"], true);
-        assert_eq!(json["enableCoreDumps"], true);
-        // Fields not exposed in CLI are omitted from the JSON
-        assert!(json.get("byocId").is_none());
-    }
-
-    #[test]
-    fn build_create_service_request_trims_tag_keys() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replica_memory_gb: None,
-            max_replica_memory_gb: None,
-            num_replicas: None,
-            min_replicas: None,
-            max_replicas: None,
-            autoscaling_mode: None,
-            idle_scaling: None,
-            idle_timeout_minutes: None,
-            ip_allow: vec![],
-            backup_id: None,
-            release_channel: None,
-            data_warehouse_id: None,
-            is_readonly: false,
-            encryption_key: None,
-            encryption_role: None,
-            enable_tde: false,
-            compliance_type: None,
-            profile: None,
-            tags: vec![" env =prod".to_string()],
-            enable_endpoints: vec![],
-            disable_endpoints: vec![],
-            private_preview_terms_checked: false,
-            enable_core_dumps: None,
-            org_id: None,
-        };
-
-        let request = build_create_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["tags"][0]["key"], "env");
-        assert_eq!(json["tags"][0]["value"], "prod");
-    }
-
-    #[test]
-    fn build_create_service_request_rejects_empty_tag_keys() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replica_memory_gb: None,
-            max_replica_memory_gb: None,
-            num_replicas: None,
-            min_replicas: None,
-            max_replicas: None,
-            autoscaling_mode: None,
-            idle_scaling: None,
-            idle_timeout_minutes: None,
-            ip_allow: vec![],
-            backup_id: None,
-            release_channel: None,
-            data_warehouse_id: None,
-            is_readonly: false,
-            encryption_key: None,
-            encryption_role: None,
-            enable_tde: false,
-            compliance_type: None,
-            profile: None,
-            tags: vec!["=prod".to_string()],
-            enable_endpoints: vec![],
-            disable_endpoints: vec![],
-            private_preview_terms_checked: false,
-            enable_core_dumps: None,
-            org_id: None,
-        };
-
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert_eq!(
-            err.to_string(),
-            "invalid tag '=prod': tag key cannot be empty"
-        );
-    }
-
-    #[test]
-    fn build_update_service_request_supports_patch_fields() {
-        let opts = ServiceUpdateOptions {
-            name: Some("updated".to_string()),
-            add_ip_allow: vec!["10.0.0.0/8".to_string()],
-            remove_ip_allow: vec!["0.0.0.0/0".to_string()],
-            add_private_endpoint_ids: vec!["pe-1".to_string()],
-            remove_private_endpoint_ids: vec!["pe-2".to_string()],
-            release_channel: Some("default".to_string()),
-            enable_endpoints: vec!["mysql".to_string()],
-            disable_endpoints: vec![],
-            transparent_data_encryption_key_id: Some("tde-1".to_string()),
-            add_tags: vec!["env=staging".to_string()],
-            remove_tags: vec!["old=tag".to_string()],
-            enable_core_dumps: Some(false),
-            org_id: None,
-        };
-
-        let request = build_update_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["ipAccessList"]["add"][0]["source"], "10.0.0.0/8");
-        assert_eq!(json["ipAccessList"]["remove"][0]["source"], "0.0.0.0/0");
-        assert_eq!(json["privateEndpointIds"]["add"][0], "pe-1");
-        assert_eq!(json["privateEndpointIds"]["remove"][0], "pe-2");
-        assert!(json["tags"].is_object());
-        assert_eq!(json["tags"]["add"][0]["key"], "env");
-        assert_eq!(json["tags"]["remove"][0]["key"], "old");
-        assert_eq!(json["transparentDataEncryptionKeyId"], "tde-1");
-        assert_eq!(json["enableCoreDumps"], false);
-    }
-
-    #[test]
-    fn build_update_service_request_rejects_empty_tag_keys() {
-        let opts = ServiceUpdateOptions {
-            name: None,
-            add_ip_allow: vec![],
-            remove_ip_allow: vec![],
-            add_private_endpoint_ids: vec![],
-            remove_private_endpoint_ids: vec![],
-            release_channel: None,
-            enable_endpoints: vec![],
-            disable_endpoints: vec![],
-            transparent_data_encryption_key_id: None,
-            add_tags: vec![" =prod".to_string()],
-            remove_tags: vec![],
-            enable_core_dumps: None,
-            org_id: None,
-        };
-
-        let err = build_update_service_request(&opts).unwrap_err();
-        assert_eq!(
-            err.to_string(),
-            "invalid tag ' =prod': tag key cannot be empty"
-        );
-    }
-
-    // Regression tests: invalid enum values must be rejected by build_* functions
-    // before any network call (resolve_org_id). See issue #101.
-
-    #[test]
-    fn build_create_service_request_rejects_invalid_provider() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "awss".to_string(),
-            region: "us-east-1".to_string(),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("awss"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_create_service_request_rejects_invalid_region() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-99".to_string(),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("us-east-99"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_create_service_request_rejects_invalid_release_channel() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            release_channel: Some("turbo".to_string()),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("turbo"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_create_service_request_horizontal_autoscaling_on_wire() {
-        // Maximal: explicit --autoscaling-mode horizontal + min/max replicas.
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replicas: Some(2),
-            max_replicas: Some(8),
-            autoscaling_mode: Some("horizontal".to_string()),
-            ..Default::default()
-        };
-        let request = build_create_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["autoscalingMode"], "horizontal");
-        assert_eq!(json["minReplicas"], 2);
-        assert_eq!(json["maxReplicas"], 8);
-        assert!(json["minReplicas"].is_i64());
-        assert!(json["maxReplicas"].is_i64());
-        // Vertical fields stay absent.
-        assert!(json.get("numReplicas").is_none());
-        assert!(json.get("minReplicaMemoryGb").is_none());
-        assert!(json.get("maxReplicaMemoryGb").is_none());
-    }
-
-    #[test]
-    fn build_create_service_request_replica_pair_without_mode_omits_mode() {
-        // No explicit --autoscaling-mode: the replica pair passes through with
-        // the mode absent. The API resolves an omitted mode itself — an equal
-        // band is accepted as a vertical fixed replica count without the
-        // horizontal entitlement, so the CLI must not inject "horizontal".
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replicas: Some(1),
-            max_replicas: Some(4),
-            ..Default::default()
-        };
-        let request = build_create_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert!(json.get("autoscalingMode").is_none());
-        assert_eq!(json["minReplicas"], 1);
-        assert_eq!(json["maxReplicas"], 4);
-    }
-
-    #[test]
-    fn build_create_service_request_vertical_omits_horizontal_fields() {
-        // Minimal: vertical-only usage leaves autoscalingMode/replicas absent.
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            num_replicas: Some(3),
-            min_replica_memory_gb: Some(24),
-            max_replica_memory_gb: Some(48),
-            ..Default::default()
-        };
-        let request = build_create_service_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert!(json.get("autoscalingMode").is_none());
-        assert!(json.get("minReplicas").is_none());
-        assert!(json.get("maxReplicas").is_none());
-        assert_eq!(json["numReplicas"], 3);
-        assert!(json["numReplicas"].is_i64());
-    }
-
-    #[test]
-    fn build_create_service_request_rejects_min_without_max_replicas() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replicas: Some(2),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("--min-replicas"),
-            "error should guide the user: {}",
-            err
-        );
-
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            max_replicas: Some(8),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("--min-replicas"),
-            "error should guide the user: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_create_service_request_rejects_invalid_autoscaling_mode() {
-        let opts = CreateServiceOptions {
-            name: "svc".to_string(),
-            provider: "aws".to_string(),
-            region: "us-east-1".to_string(),
-            min_replicas: Some(2),
-            max_replicas: Some(8),
-            autoscaling_mode: Some("turbo".to_string()),
-            ..Default::default()
-        };
-        let err = build_create_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("turbo"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn resolve_horizontal_autoscaling_explicit_vertical_with_no_replicas() {
-        // Explicit --autoscaling-mode vertical with no replica pair → mode set,
-        // replicas absent (lets the server apply vertical defaults).
-        let resolved = resolve_horizontal_autoscaling(Some("vertical"), None, None).unwrap();
-        assert_eq!(resolved.autoscaling_mode, Some(AutoscalingMode::Vertical));
-        assert!(resolved.min_replicas.is_none());
-        assert!(resolved.max_replicas.is_none());
-    }
-
-    #[test]
-    fn build_service_scale_request_horizontal_autoscaling_on_wire() {
-        let opts = ServiceScaleOptions {
-            min_replicas: Some(2),
-            max_replicas: Some(8),
-            autoscaling_mode: Some("horizontal".to_string()),
-            ..Default::default()
-        };
-        let request = build_service_scale_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["autoscalingMode"], "horizontal");
-        assert_eq!(json["minReplicas"], 2);
-        assert_eq!(json["maxReplicas"], 8);
-        assert!(json.get("numReplicas").is_none());
-        assert!(json.get("minReplicaMemoryGb").is_none());
-        assert!(json.get("maxReplicaMemoryGb").is_none());
-    }
-
-    #[test]
-    fn build_service_scale_request_switch_to_vertical_on_wire() {
-        // Switching a horizontal service back to vertical sends the mode and
-        // the vertical fields in one request.
-        let opts = ServiceScaleOptions {
-            autoscaling_mode: Some("vertical".to_string()),
-            num_replicas: Some(3),
-            min_replica_memory_gb: Some(8),
-            max_replica_memory_gb: Some(32),
-            ..Default::default()
-        };
-        let request = build_service_scale_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["autoscalingMode"], "vertical");
-        assert_eq!(json["numReplicas"], 3);
-        assert_eq!(json["minReplicaMemoryGb"], 8.0);
-        assert_eq!(json["maxReplicaMemoryGb"], 32.0);
-        assert!(json.get("minReplicas").is_none());
-        assert!(json.get("maxReplicas").is_none());
-    }
-
-    #[test]
-    fn build_service_scale_request_switch_to_horizontal_with_memory_on_wire() {
-        // Switching to horizontal pins the equal per-replica memory the mode
-        // requires in the same request.
-        let opts = ServiceScaleOptions {
-            autoscaling_mode: Some("horizontal".to_string()),
-            min_replicas: Some(2),
-            max_replicas: Some(8),
-            min_replica_memory_gb: Some(16),
-            max_replica_memory_gb: Some(16),
-            ..Default::default()
-        };
-        let request = build_service_scale_request(&opts).unwrap();
-        let json = serde_json::to_value(&request).unwrap();
-        assert_eq!(json["autoscalingMode"], "horizontal");
-        assert_eq!(json["minReplicas"], 2);
-        assert_eq!(json["maxReplicas"], 8);
-        assert_eq!(json["minReplicaMemoryGb"], 16.0);
-        assert_eq!(json["maxReplicaMemoryGb"], 16.0);
-        assert!(json.get("numReplicas").is_none());
-    }
-
-    #[test]
-    fn build_service_scale_request_rejects_min_without_max_replicas() {
-        let opts = ServiceScaleOptions {
-            max_replicas: Some(8),
-            ..Default::default()
-        };
-        let err = build_service_scale_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("--min-replicas"),
-            "error should guide the user: {}",
-            err
-        );
-    }
 
     #[test]
     fn build_kinesis_source_rejects_out_of_range_iterator_timestamp() {
@@ -3054,43 +1013,6 @@ mod tests {
         };
         let source = build_kinesis_source(&args).unwrap();
         assert_eq!(source.timestamp, Some(1_750_000_000));
-    }
-
-    #[test]
-    fn build_update_service_request_rejects_invalid_release_channel() {
-        let opts = ServiceUpdateOptions {
-            release_channel: Some("turbo".to_string()),
-            ..Default::default()
-        };
-        let err = build_update_service_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("turbo"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_password_and_query_endpoint_requests_use_new_fields() {
-        let password_request = build_service_password_patch_request(&ServiceResetPasswordOptions {
-            new_password_hash: Some("sha256".to_string()),
-            new_double_sha1_hash: Some("sha1".to_string()),
-            org_id: None,
-        });
-        let password_json = serde_json::to_value(&password_request).unwrap();
-        assert_eq!(password_json["newPasswordHash"], "sha256");
-        assert_eq!(password_json["newDoubleSha1Hash"], "sha1");
-
-        let query_request = build_query_endpoint_create_request(&QueryEndpointCreateOptions {
-            roles: vec!["admin".to_string()],
-            open_api_keys: vec!["key-1".to_string()],
-            allowed_origins: Some("https://example.com".to_string()),
-            org_id: None,
-        });
-        let query_json = serde_json::to_value(&query_request).unwrap();
-        assert_eq!(query_json["roles"][0], "admin");
-        assert_eq!(query_json["openApiKeys"][0], "key-1");
-        assert_eq!(query_json["allowedOrigins"], "https://example.com");
     }
 
     #[test]
