+++ crates/clickhousectl/src/cloud/mod.rs
@@ -10,6 +10,7 @@ pub mod organizations;
 pub mod output;
 pub mod postgres;
 pub mod service_query;
+pub mod services;
 mod shared;
 pub mod types;
 
@@ -23,8 +24,7 @@ pub use client::{
 
 use crate::error::{Error, Result};
 use cli::{
-    ClickPipeCommands, ClickPipeCreateCommands, ClickPipeSettingsCommands, CloudArgs,
-    CloudCommands, PrivateEndpointCommands, QueryEndpointCommands, ServiceCommands,
+    ClickPipeCommands, ClickPipeCreateCommands, ClickPipeSettingsCommands, CloudArgs, CloudCommands,
 };
 
 /// Explain when a configured environment credential cannot participate in
@@ -128,253 +128,7 @@ async fn dispatch(
     match command {
         CloudCommands::Auth { .. } => unreachable!("handled above"),
         CloudCommands::Org { command } => organizations::run_org(client, command, json).await,
-        CloudCommands::Service { command } => match command {
-            ServiceCommands::List { org_id, filter } => {
-                commands::service_list(client, org_id.as_deref(), &filter, json).await
-            }
-            ServiceCommands::Get { service_id, org_id } => {
-                commands::service_get(client, &service_id, org_id.as_deref(), json).await
-            }
-            ServiceCommands::Create {
-                name,
-                provider,
-                region,
-                min_replica_memory_gb,
-                max_replica_memory_gb,
-                num_replicas,
-                min_replicas,
-                max_replicas,
-                autoscaling_mode,
-                idle_scaling,
-                idle_timeout_minutes,
-                ip_allow,
-                backup_id,
-                release_channel,
-                data_warehouse_id,
-                readonly,
-                encryption_key,
-                encryption_role,
-                enable_tde,
-                compliance_type,
-                profile,
-                tag,
-                enable_endpoint,
-                disable_endpoint,
-                private_preview_terms_checked,
-                enable_core_dumps,
-                org_id,
-            } => {
-                let opts = commands::CreateServiceOptions {
-                    name,
-                    provider,
-                    region,
-                    min_replica_memory_gb,
-                    max_replica_memory_gb,
-                    num_replicas,
-                    min_replicas,
-                    max_replicas,
-                    autoscaling_mode,
-                    idle_scaling,
-                    idle_timeout_minutes,
-                    ip_allow,
-                    backup_id,
-                    release_channel,
-                    data_warehouse_id,
-                    is_readonly: readonly,
-                    encryption_key,
-                    encryption_role,
-                    enable_tde,
-                    compliance_type,
-                    profile,
-                    tags: tag,
-                    enable_endpoints: enable_endpoint,
-                    disable_endpoints: disable_endpoint,
-                    private_preview_terms_checked,
-                    enable_core_dumps,
-                    org_id,
-                };
-                commands::service_create(client, opts, json).await
-            }
-            ServiceCommands::Delete {
-                service_id,
-                force,
-                org_id,
-            } => {
-                commands::service_delete(client, &service_id, force, org_id.as_deref(), json).await
-            }
-            ServiceCommands::Start { service_id, org_id } => {
-                commands::service_start(client, &service_id, org_id.as_deref(), json).await
-            }
-            ServiceCommands::Stop { service_id, org_id } => {
-                commands::service_stop(client, &service_id, org_id.as_deref(), json).await
-            }
-            ServiceCommands::Update {
-                service_id,
-                name,
-                add_ip_allow,
-                remove_ip_allow,
-                add_private_endpoint_id,
-                remove_private_endpoint_id,
-                release_channel,
-                enable_endpoint,
-                disable_endpoint,
-                transparent_data_encryption_key_id,
-                add_tag,
-                remove_tag,
-                enable_core_dumps,
-                org_id,
-            } => {
-                let opts = commands::ServiceUpdateOptions {
-                    name,
-                    add_ip_allow,
-                    remove_ip_allow,
-                    add_private_endpoint_ids: add_private_endpoint_id,
-                    remove_private_endpoint_ids: remove_private_endpoint_id,
-                    release_channel,
-                    enable_endpoints: enable_endpoint,
-                    disable_endpoints: disable_endpoint,
-                    transparent_data_encryption_key_id,
-                    add_tags: add_tag,
-                    remove_tags: remove_tag,
-                    enable_core_dumps,
-                    org_id,
-                };
-                commands::service_update(client, &service_id, opts, json).await
-            }
-            ServiceCommands::Scale {
-                service_id,
-                min_replica_memory_gb,
-                max_replica_memory_gb,
-                num_replicas,
-                min_replicas,
-                max_replicas,
-                autoscaling_mode,
-                idle_scaling,
-                idle_timeout_minutes,
-                org_id,
-            } => {
-                commands::service_scale(
-                    client,
-                    &service_id,
-                    commands::ServiceScaleOptions {
-                        min_replica_memory_gb,
-                        max_replica_memory_gb,
-                        num_replicas,
-                        min_replicas,
-                        max_replicas,
-                        autoscaling_mode,
-                        idle_scaling,
-                        idle_timeout_minutes,
-                        org_id,
-                    },
-                    json,
-                )
-                .await
-            }
-            ServiceCommands::ResetPassword {
-                service_id,
-                new_password_hash,
-                new_double_sha1_hash,
-                org_id,
-            } => {
-                let opts = commands::ServiceResetPasswordOptions {
-                    new_password_hash,
-                    new_double_sha1_hash,
-                    org_id,
-                };
-                commands::service_reset_password(client, &service_id, opts, json).await
-            }
-            ServiceCommands::QueryEndpoint { command } => match command {
-                QueryEndpointCommands::Get { service_id, org_id } => {
-                    commands::query_endpoint_get(client, &service_id, org_id.as_deref(), json).await
-                }
-                QueryEndpointCommands::Create {
-                    service_id,
-                    role,
-                    open_api_key,
-                    allowed_origins,
-                    org_id,
-                } => {
-                    let opts = commands::QueryEndpointCreateOptions {
-                        roles: role,
-                        open_api_keys: open_api_key,
-                        allowed_origins,
-                        org_id,
-                    };
-                    commands::query_endpoint_create(client, &service_id, opts, json).await
-                }
-                QueryEndpointCommands::Delete { service_id, org_id } => {
-                    commands::query_endpoint_delete(client, &service_id, org_id.as_deref(), json)
-                        .await
-                }
-            },
-            ServiceCommands::PrivateEndpoint { command } => match command {
-                PrivateEndpointCommands::Create {
-                    service_id,
-                    endpoint_id,
-                    description,
-                    org_id,
-                } => {
-                    commands::private_endpoint_create(
-                        client,
-                        &service_id,
-                        &endpoint_id,
-                        description.as_deref(),
-                        org_id.as_deref(),
-                        json,
-                    )
-                    .await
-                }
-                PrivateEndpointCommands::GetConfig { service_id, org_id } => {
-                    commands::private_endpoint_get_config(
-                        client,
-                        &service_id,
-                        org_id.as_deref(),
-                        json,
-                    )
-                    .await
-                }
-            },
-            ServiceCommands::BackupConfig { command } => {
-                backups::run_config(client, command, json).await
-            }
-            ServiceCommands::Prometheus {
-                service_id,
-                org_id,
-                filtered_metrics,
-            } => {
-                commands::service_prometheus(
-                    client,
-                    &service_id,
-                    org_id.as_deref(),
-                    filtered_metrics,
-                )
-                .await
-            }
-            ServiceCommands::Query {
-                name,
-                id,
-                query,
-                queries_file,
-                database,
-                format,
-                org_id,
-                no_auto_enable,
-            } => {
-                let opts = commands::ServiceQueryOptions {
-                    name,
-                    id,
-                    query,
-                    queries_file,
-                    database,
-                    format,
-                    json,
-                    org_id,
-                    no_auto_enable,
-                };
-                commands::service_query(client, opts).await
-            }
-        },
+        CloudCommands::Service { command } => services::run(client, command, json).await,
         CloudCommands::Member { command } => organizations::run_member(client, command, json).await,
         CloudCommands::Invitation { command } => {
             organizations::run_invitation(client, command, json).await
